Cardano.Ledger.Alonzo.BlockBody.Internal

Description

Provides BlockBody internals

Warning

This module is considered internal.

The contents of this module may change in any way whatsoever and without any warning between minor versions of this package.

Documentation

data AlonzoBlockBody era

Constructors

AlonzoBlockBodyInternal

Fields

abbTxs ∷ !(StrictSeq (Tx era))

abbHash ∷ Hash HASH EraIndependentBlockBody

Memoized hash to avoid recomputation. Lazy on purpose.

abbTxsBodyBytes ∷ ByteString

Bytes encoding Seq (TxBody era)

abbTxsWitsBytes ∷ ByteString

Bytes encoding Seq (TxWits era)

abbTxsAuxDataBytes ∷ ByteString

Bytes encoding a TxAuxData). Missing indices have SNothing for metadata

abbTxsIsValidBytes ∷ ByteString

Bytes representing a set of integers. These are the indices of transactions with isValid == False.

Bundled Patterns

pattern AlonzoBlockBody ∷ ∀ era. (AlonzoEraTx era, SafeToHash (TxWits era)) ⇒ StrictSeq (Tx era) → AlonzoBlockBody era

hashAlonzoSegWits

Arguments

∷ ByteString

→ ByteString
Bytes for transaction bodies

→ ByteString
Bytes for transaction witnesses

→ ByteString
Bytes for transaction auxiliary datas

→ Hash HASH EraIndependentBlockBody
Bytes for transaction isValid flags

alignedValidFlags ∷ Int → [Int] → Seq IsValid

Given the number of transactions, and the set of indices for which these transactions do not validate, create an aligned sequence of IsValid flags.

This function operates much as the inverse of nonValidatingIndices.

mkBasicBlockBodyAlonzo ∷ (SafeToHash (TxWits era), BlockBody era ~ AlonzoBlockBody era, AlonzoEraTx era) ⇒ BlockBody era

txSeqBlockBodyAlonzoL ∷ (SafeToHash (TxWits era), BlockBody era ~ AlonzoBlockBody era, AlonzoEraTx era) ⇒ Lens' (BlockBody era) (StrictSeq (Tx era))

Orphan instances

EraBlockBody AlonzoEra

Instance details

Associated Types

type BlockBody AlonzoEra = (r ∷ Type)

Methods

mkBasicBlockBody ∷ BlockBody AlonzoEra

txSeqBlockBodyL ∷ Lens' (BlockBody AlonzoEra) (StrictSeq (Tx AlonzoEra))

fromTxSeq ∷ BlockBody AlonzoEra → StrictSeq (Tx AlonzoEra)

toTxSeq ∷ StrictSeq (Tx AlonzoEra) → BlockBody AlonzoEra

hashBlockBody ∷ BlockBody AlonzoEra → Hash HASH EraIndependentBlockBody

hashTxSeq ∷ BlockBody AlonzoEra → Hash HASH EraIndependentBlockBody

numSegComponents ∷ Word64