Cardano.Ledger.Alonzo.Core

Documentation

class (EraTx era, AlonzoEraTxBody era, AlonzoEraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTx era where

Methods

isValidTxL ∷ Lens' (Tx era) IsValid

newtype IsValid

Tag indicating whether non-native scripts in this transaction are expected to validate. This is added by the block creator when constructing the block.

Constructors

IsValid Bool

class (AlonzoEraPParams era, EraTxOut era) ⇒ AlonzoEraTxOut era where

Methods

dataHashTxOutL ∷ Lens' (TxOut era) (StrictMaybe DataHash)

datumTxOutF ∷ SimpleGetter (TxOut era) (Datum era)

class (EraScript era, Eq (PlutusScript era), Ord (PlutusScript era), Show (PlutusScript era), NoThunks (PlutusScript era), NFData (PlutusScript era), SafeToHash (PlutusScript era), Eq (PlutusPurpose AsItem era), Show (PlutusPurpose AsItem era), EncCBOR (PlutusPurpose AsItem era), DecCBOR (PlutusPurpose AsItem era), NoThunks (PlutusPurpose AsItem era), NFData (PlutusPurpose AsItem era), Eq (PlutusPurpose AsIx era), Ord (PlutusPurpose AsIx era), Show (PlutusPurpose AsIx era), EncCBOR (PlutusPurpose AsIx era), DecCBOR (PlutusPurpose AsIx era), EncCBORGroup (PlutusPurpose AsIx era), DecCBORGroup (PlutusPurpose AsIx era), NoThunks (PlutusPurpose AsIx era), NFData (PlutusPurpose AsIx era), Eq (PlutusPurpose AsIxItem era), Show (PlutusPurpose AsIxItem era), NoThunks (PlutusPurpose AsIxItem era), NFData (PlutusPurpose AsIxItem era), AllegraEraScript era) ⇒ AlonzoEraScript era where

Minimal complete definition

eraMaxLanguage, mkPlutusScript, withPlutusScript, hoistPlutusPurpose, mkSpendingPurpose, toSpendingPurpose, mkMintingPurpose, toMintingPurpose, mkCertifyingPurpose, toCertifyingPurpose, mkRewardingPurpose, toRewardingPurpose, upgradePlutusPurposeAsIx

Associated Types

data PlutusScript era ∷ Type

type PlutusPurpose (f ∷ Type → Type → Type) era = (r ∷ Type) | r → era

Methods

eraMaxLanguage ∷ Language

Highest supported Plutus language version for this era.

toPlutusScript ∷ Script era → Maybe (PlutusScript era)

Attempt to extract a AlonzoEraScript from a wrapper type family Script. Whenevr Script is a native script Nothing will be returned

default toPlutusScript ∷ Script era ~ AlonzoScript era ⇒ Script era → Maybe (PlutusScript era)

fromPlutusScript ∷ PlutusScript era → Script era

Convert a AlonzoEraScript to a wrapper type family Script

default fromPlutusScript ∷ Script era ~ AlonzoScript era ⇒ PlutusScript era → Script era

mkPlutusScript ∷ (PlutusLanguage l, MonadFail m) ⇒ Plutus l → m (PlutusScript era)

Returns Nothing, whenver plutus language is not supported for this era.

withPlutusScript ∷ PlutusScript era → (∀ l. PlutusLanguage l ⇒ Plutus l → a) → a

Give a AlonzoEraScript apply a function that can handle Plutus scripts of all known versions.

hoistPlutusPurpose ∷ (∀ ix it. g ix it → f ix it) → PlutusPurpose g era → PlutusPurpose f era

mkSpendingPurpose ∷ f Word32 TxIn → PlutusPurpose f era

toSpendingPurpose ∷ PlutusPurpose f era → Maybe (f Word32 TxIn)

mkMintingPurpose ∷ f Word32 PolicyID → PlutusPurpose f era

toMintingPurpose ∷ PlutusPurpose f era → Maybe (f Word32 PolicyID)

mkCertifyingPurpose ∷ f Word32 (TxCert era) → PlutusPurpose f era

toCertifyingPurpose ∷ PlutusPurpose f era → Maybe (f Word32 (TxCert era))

mkRewardingPurpose ∷ f Word32 RewardAccount → PlutusPurpose f era

toRewardingPurpose ∷ PlutusPurpose f era → Maybe (f Word32 RewardAccount)

upgradePlutusPurposeAsIx ∷ AlonzoEraScript (PreviousEra era) ⇒ PlutusPurpose AsIx (PreviousEra era) → PlutusPurpose AsIx era

newtype AsIx ix it

Constructors

AsIx

Fields

unAsIx ∷ ix

newtype AsItem ix it

Constructors

AsItem

Fields

unAsItem ∷ it

data AsIxItem ix it

Constructors

AsIxItem

Fields

asIndex ∷ !ix

asItem ∷ !it

pattern SpendingPurpose ∷ AlonzoEraScript era ⇒ f Word32 TxIn → PlutusPurpose f era

pattern MintingPurpose ∷ AlonzoEraScript era ⇒ f Word32 PolicyID → PlutusPurpose f era

pattern CertifyingPurpose ∷ AlonzoEraScript era ⇒ f Word32 (TxCert era) → PlutusPurpose f era

pattern RewardingPurpose ∷ AlonzoEraScript era ⇒ f Word32 RewardAccount → PlutusPurpose f era

type ScriptIntegrityHash = SafeHash EraIndependentScriptIntegrity

class (MaryEraTxBody era, AlonzoEraTxOut era) ⇒ AlonzoEraTxBody era where

Minimal complete definition

collateralInputsTxBodyL, reqSignerHashesTxBodyL, scriptIntegrityHashTxBodyL, networkIdTxBodyL, redeemerPointer, redeemerPointerInverse

Methods

collateralInputsTxBodyL ∷ Lens' (TxBody era) (Set TxIn)

reqSignerHashesTxBodyL ∷ AtMostEra "Conway" era ⇒ Lens' (TxBody era) (Set (KeyHash 'Witness))

reqSignerHashesTxBodyG ∷ SimpleGetter (TxBody era) (Set (KeyHash Witness))

default reqSignerHashesTxBodyG ∷ AtMostEra "Conway" era ⇒ SimpleGetter (TxBody era) (Set (KeyHash Witness))

scriptIntegrityHashTxBodyL ∷ Lens' (TxBody era) (StrictMaybe ScriptIntegrityHash)

networkIdTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Network)

redeemerPointer ∷ TxBody era → PlutusPurpose AsItem era → StrictMaybe (PlutusPurpose AsIx era)

This function is called rdptr in the spec. Given a TxBody and a plutus purpose with an item, we should be able to find the plutus purpose as in index

redeemerPointerInverse ∷ TxBody era → PlutusPurpose AsIx era → StrictMaybe (PlutusPurpose AsIxItem era)

This is an inverse of redeemerPointer. Given purpose as an index return it as an item.

class (EraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTxWits era where

Methods

datsTxWitsL ∷ Lens' (TxWits era) (TxDats era)

rdmrsTxWitsL ∷ Lens' (TxWits era) (Redeemers era)

class EraPParams era ⇒ AlonzoEraPParams era

Minimal complete definition

hkdCoinsPerUTxOWordL, hkdCostModelsL, hkdPricesL, hkdMaxTxExUnitsL, hkdMaxBlockExUnitsL, hkdMaxValSizeL, hkdCollateralPercentageL, hkdMaxCollateralInputsL

newtype CoinPerWord

Constructors

CoinPerWord

Fields

unCoinPerWord ∷ Coin

ppCoinsPerUTxOWordL ∷ ∀ era. (AlonzoEraPParams era, ExactEra AlonzoEra era) ⇒ Lens' (PParams era) CoinPerWord

ppCostModelsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) CostModels

ppPricesL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) Prices

ppMaxTxExUnitsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) ExUnits

ppMaxBlockExUnitsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) ExUnits

ppMaxValSizeL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppCollateralPercentageL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppMaxCollateralInputsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppuCoinsPerUTxOWordL ∷ ∀ era. (AlonzoEraPParams era, ExactEra AlonzoEra era) ⇒ Lens' (PParamsUpdate era) (StrictMaybe CoinPerWord)

ppuCostModelsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe CostModels)

ppuPricesL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Prices)

ppuMaxTxExUnitsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe ExUnits)

ppuMaxBlockExUnitsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe ExUnits)

ppuMaxValSizeL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

ppuCollateralPercentageL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

ppuMaxCollateralInputsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

class AllegraEraTxAuxData era ⇒ AlonzoEraTxAuxData era where

Methods

plutusScriptsTxAuxDataL ∷ Lens' (TxAuxData era) (Map Language (NonEmpty PlutusBinary))

module Cardano.Ledger.Mary.Core