Cardano.Ledger.Alonzo.Data

Documentation

data Data era where

Bundled Patterns

pattern Data :: Data -> Data era

type DataHash crypto = SafeHash crypto EraIndependentData

hashData :: Era era => Data era -> DataHash (Crypto era)

getPlutusData :: Data era -> Data

dataHashSize :: StrictMaybe (DataHash c) -> Integer

data BinaryData era

Inlined data must be stored in the most compact form because it contributes to the memory overhead of the ledger state. Constructor is intentionally not exported, in order to prevent invalid creation of data from arbitrary binary data. Use makeBinaryData for smart construction.

hashBinaryData :: Era era => BinaryData era -> DataHash (Crypto era)

makeBinaryData :: ShortByteString -> Either String (BinaryData era)

binaryDataToData :: BinaryData era -> Data era

It is safe to convert BinaryData to Data because the only way to construct BinaryData is thorugh smart constructor makeBinaryData that takes care of verification.

dataToBinaryData :: Data era -> BinaryData era

decodeBinaryData :: BinaryData era -> Either DecoderError (Data era)

data Datum era

Datum can be described by a either a data hash or binary data, but not both. It can also be neither one of them.

Constructors

NoDatum

DatumHash !(DataHash (Crypto era))

Datum !(BinaryData era)

datumDataHash :: Era era => Datum era -> StrictMaybe (DataHash (Crypto era))

Get the Hash of the datum.

data AuxiliaryData era where

Bundled Patterns

pattern AuxiliaryData :: (Era era, ToCBOR (Script era), Script era ~ Script era, Ord (Script era)) => Map Word64 Metadatum -> StrictSeq (Script era) -> AuxiliaryData era

pattern AuxiliaryData' :: Map Word64 Metadatum -> StrictSeq (Script era) -> AuxiliaryData era

newtype AuxiliaryDataHash crypto

Constructors

AuxiliaryDataHash

Fields

unsafeAuxiliaryDataHash :: SafeHash crypto EraIndependentAuxiliaryData

Orphan instances

ToCBOR Data

Instance details

Methods

toCBOR :: Data -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy Data -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [Data] -> Size

NoThunks Data

Instance details

Methods

noThunks :: Context -> Data -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> Data -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy Data -> String

FromCBOR (Annotator Data)

Instance details

Methods

fromCBOR :: Decoder s (Annotator Data)

label :: Proxy (Annotator Data) -> Text

Crypto c => HeapWords (StrictMaybe (DataHash c))

Instance details

Methods

heapWords :: StrictMaybe (DataHash c) -> Int