Cardano.Ledger.Alonzo.PParams

Description

This module contains just the type of protocol parameters.

Era Agnostic

class EraPParams era ⇒ AlonzoEraPParams era where

Methods

hkdCoinsPerUTxOWordL ∷ (HKDFunctor f, ExactEra AlonzoEra era) ⇒ Lens' (PParamsHKD f era) (HKD f CoinPerWord)

hkdCostModelsL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f CostModels)

hkdPricesL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Prices)

hkdMaxTxExUnitsL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f ExUnits)

hkdMaxBlockExUnitsL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f ExUnits)

hkdMaxValSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Natural)

hkdCollateralPercentageL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Natural)

hkdMaxCollateralInputsL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Natural)

ppCoinsPerUTxOWordL ∷ ∀ era. (AlonzoEraPParams era, ExactEra AlonzoEra era) ⇒ Lens' (PParams era) CoinPerWord

ppCostModelsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) CostModels

ppPricesL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) Prices

ppMaxTxExUnitsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) ExUnits

ppMaxBlockExUnitsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) ExUnits

ppMaxValSizeL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppCollateralPercentageL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppMaxCollateralInputsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppuCoinsPerUTxOWordL ∷ ∀ era. (AlonzoEraPParams era, ExactEra AlonzoEra era) ⇒ Lens' (PParamsUpdate era) (StrictMaybe CoinPerWord)

ppuCostModelsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe CostModels)

ppuPricesL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Prices)

ppuMaxTxExUnitsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe ExUnits)

ppuMaxBlockExUnitsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe ExUnits)

ppuMaxValSizeL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

ppuCollateralPercentageL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

ppuMaxCollateralInputsL ∷ ∀ era. AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

Alonzo specific

data AlonzoPParams f era

Protocol parameters. Shelley parameters + additional ones

Constructors

AlonzoPParams

Fields

appMinFeeA ∷ !(HKD f Coin)

The linear factor for the minimum fee calculation

appMinFeeB ∷ !(HKD f Coin)

The constant factor for the minimum fee calculation

appMaxBBSize ∷ !(HKD f Word32)

Maximal block body size

appMaxTxSize ∷ !(HKD f Word32)

Maximal transaction size

appMaxBHSize ∷ !(HKD f Word16)

Maximal block header size

appKeyDeposit ∷ !(HKD f Coin)

The amount of a key registration deposit

appPoolDeposit ∷ !(HKD f (CompactForm Coin))

The amount of a pool registration deposit

appEMax ∷ !(HKD f EpochInterval)

Maximum number of epochs in the future a pool retirement is allowed to be scheduled for.

appNOpt ∷ !(HKD f Word16)

Desired number of pools

appA0 ∷ !(HKD f NonNegativeInterval)

Pool influence

appRho ∷ !(HKD f UnitInterval)

Monetary expansion

appTau ∷ !(HKD f UnitInterval)

Treasury expansion

appD ∷ !(HKD f UnitInterval)

Decentralization parameter. Note that the scale is inverted here - a value of 0 indicates full decentralisation, where 1 indicates full federalisation.

appExtraEntropy ∷ !(HKD f Nonce)

Extra entropy

appProtocolVersion ∷ !(HKD f ProtVer)

Protocol version

appMinPoolCost ∷ !(HKD f Coin)

Minimum Stake Pool Cost

appCoinsPerUTxOWord ∷ !(HKD f CoinPerWord)

Cost in lovelace per word (8 bytes) of UTxO storage (instead of appMinUTxOValue)

appCostModels ∷ !(HKD f CostModels)

Cost models for non-native script languages

appPrices ∷ !(HKD f Prices)

Prices of execution units (for non-native script languages)

appMaxTxExUnits ∷ !(HKD f OrdExUnits)

Max total script execution resources units allowed per tx

appMaxBlockExUnits ∷ !(HKD f OrdExUnits)

Max total script execution resources units allowed per block

appMaxValSize ∷ !(HKD f Natural)

Max size of a Value in an output

appCollateralPercentage ∷ !(HKD f Natural)

Percentage of the txfee which must be provided as collateral when including non-native scripts.

appMaxCollateralInputs ∷ !(HKD f Natural)

Maximum number of collateral inputs allowed in a transaction

data UpgradeAlonzoPParams f

Parameters that were added in Alonzo

Constructors

UpgradeAlonzoPParams

Fields

uappCoinsPerUTxOWord ∷ !(HKD f CoinPerWord)

uappCostModels ∷ !(HKD f CostModels)

uappPrices ∷ !(HKD f Prices)

uappMaxTxExUnits ∷ !(HKD f ExUnits)

uappMaxBlockExUnits ∷ !(HKD f ExUnits)

uappMaxValSize ∷ !(HKD f Natural)

uappCollateralPercentage ∷ !(HKD f Natural)

uappMaxCollateralInputs ∷ !(HKD f Natural)

newtype DowngradeAlonzoPParams f

Parameters that were removed in Alonzo

Constructors

DowngradeAlonzoPParams

Fields

dappMinUTxOValue ∷ HKD f Coin

emptyAlonzoPParams ∷ ∀ era. Era era ⇒ AlonzoPParams Identity era

Returns a basic "empty" PParams structure with all zero values.

emptyAlonzoPParamsUpdate ∷ AlonzoPParams StrictMaybe era

upgradeAlonzoPParams ∷ ∀ f era1 era2. HKDFunctor f ⇒ UpgradeAlonzoPParams f → ShelleyPParams f era1 → AlonzoPParams f era2

Given the missing pieces, turn a ShelleyPParams into a AlonzoPParams

downgradeAlonzoPParams ∷ DowngradeAlonzoPParams f → AlonzoPParams f era2 → ShelleyPParams f era1

Turn an AlonzoPParams into a ShelleyParams

getLanguageView ∷ AlonzoEraPParams era ⇒ PParams era → Language → LangDepView

data LangDepView

Constructors

LangDepView

Fields

tag ∷ ByteString

params ∷ ByteString

encodeLangViews ∷ Set LangDepView → Encoding

newtype OrdExUnits

This is a helper type that allows us to define an Ord instance for executions units without affecting the ExUnits type. This is needed in order to derive an Ord instance` for PParams. This is just a helper type and should not be used directly. Both lenses that operate on TxExUnits and BlockExUnits use the ExUnits type, not this one.

Constructors

OrdExUnits

Fields

unOrdExUnits ∷ ExUnits

newtype CoinPerWord

Constructors

CoinPerWord

Fields

unCoinPerWord ∷ Coin

PParam

ppCollateralPercentage ∷ AlonzoEraPParams era ⇒ PParam era

ppCostModels ∷ AlonzoEraPParams era ⇒ PParam era

ppMaxBlockExUnits ∷ AlonzoEraPParams era ⇒ PParam era

ppMaxCollateralInputs ∷ AlonzoEraPParams era ⇒ PParam era

ppMaxTxExUnits ∷ AlonzoEraPParams era ⇒ PParam era

ppMaxValSize ∷ AlonzoEraPParams era ⇒ PParam era

ppPrices ∷ AlonzoEraPParams era ⇒ PParam era

Orphan instances

EraPParams AlonzoEra

Instance details

Associated Types

type PParamsHKD f AlonzoEra = (r ∷ Type)

type UpgradePParams f AlonzoEra

type DowngradePParams f AlonzoEra

Methods

applyPPUpdates ∷ PParams AlonzoEra → PParamsUpdate AlonzoEra → PParams AlonzoEra

emptyPParamsIdentity ∷ PParamsHKD Identity AlonzoEra

emptyPParamsStrictMaybe ∷ PParamsHKD StrictMaybe AlonzoEra

emptyUpgradePParamsUpdate ∷ UpgradePParams StrictMaybe AlonzoEra

upgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDApplicative f, EraPParams (PreviousEra AlonzoEra)) ⇒ UpgradePParams f AlonzoEra → PParamsHKD f (PreviousEra AlonzoEra) → PParamsHKD f AlonzoEra

downgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, EraPParams (PreviousEra AlonzoEra)) ⇒ DowngradePParams f AlonzoEra → PParamsHKD f AlonzoEra → PParamsHKD f (PreviousEra AlonzoEra)

hkdMinFeeAL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f Coin)

hkdMinFeeBL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f Coin)

hkdMaxBBSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f Word32)

hkdMaxTxSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f Word32)

hkdMaxBHSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f Word16)

hkdKeyDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f Coin)

hkdPoolDepositCompactL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f (CompactForm Coin))

hkdEMaxL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f EpochInterval)

hkdNOptL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f Word16)

hkdA0L ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f NonNegativeInterval)

hkdRhoL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f UnitInterval)

hkdTauL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f UnitInterval)

hkdDL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Alonzo" AlonzoEra) ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f UnitInterval)

ppDG ∷ SimpleGetter (PParams AlonzoEra) UnitInterval

hkdExtraEntropyL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Alonzo" AlonzoEra) ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f Nonce)

hkdProtocolVersionL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Babbage" AlonzoEra) ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f ProtVer)

ppProtocolVersionL ∷ Lens' (PParams AlonzoEra) ProtVer

ppuProtocolVersionL ∷ Lens' (PParamsUpdate AlonzoEra) (StrictMaybe ProtVer)

hkdMinUTxOValueL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, AtMostEra "Mary" AlonzoEra) ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f Coin)

hkdMinPoolCostL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f AlonzoEra) (HKD f Coin)

eraPParams ∷ [PParam AlonzoEra]

EraGov AlonzoEra

Instance details

Associated Types

type GovState AlonzoEra = (r ∷ Type)

Methods

emptyGovState ∷ GovState AlonzoEra

curPParamsGovStateL ∷ Lens' (GovState AlonzoEra) (PParams AlonzoEra)

prevPParamsGovStateL ∷ Lens' (GovState AlonzoEra) (PParams AlonzoEra)

futurePParamsGovStateL ∷ Lens' (GovState AlonzoEra) (FuturePParams AlonzoEra)

obligationGovState ∷ GovState AlonzoEra → Obligations