Cardano.Ledger.Alonzo.Plutus.Context

Documentation

data LedgerTxInfo era

All information that is necessary from the ledger to construct Plutus' TxInfo.

Constructors

LedgerTxInfo

Fields

ltiProtVer ∷ !ProtVer

ltiEpochInfo ∷ !(EpochInfo (Either Text))

ltiSystemStart ∷ !SystemStart

ltiUTxO ∷ !(UTxO era)

ltiTx ∷ !(Tx era)

class (PlutusLanguage l, EraPlutusContext era) ⇒ EraPlutusTxInfo (l ∷ Language) era where

Methods

toPlutusTxCert ∷ proxy l → ProtVer → TxCert era → Either (ContextError era) (PlutusTxCert l)

toPlutusScriptPurpose ∷ proxy l → ProtVer → PlutusPurpose AsIxItem era → Either (ContextError era) (PlutusScriptPurpose l)

toPlutusTxInfo ∷ proxy l → LedgerTxInfo era → Either (ContextError era) (PlutusTxInfo l)

toPlutusArgs ∷ proxy l → ProtVer → PlutusTxInfo l → PlutusPurpose AsIxItem era → Maybe (Data era) → Data era → Either (ContextError era) (PlutusArgs l)

toPlutusTxInInfo ∷ proxy l → UTxO era → TxIn → Either (ContextError era) (PlutusTxInInfo era l)

class (AlonzoEraScript era, Eq (ContextError era), Show (ContextError era), NFData (ContextError era), NoThunks (ContextError era), EncCBOR (ContextError era), DecCBOR (ContextError era), ToJSON (ContextError era)) ⇒ EraPlutusContext era where

Associated Types

type ContextError era = (r ∷ Type) | r → era

data TxInfoResult era ∷ Type

This data type family is used to memoize the results of toPlutusTxInfo, so the outcome can be shared between execution of different scripts with the same language version.

Methods

mkSupportedLanguage ∷ Language → Maybe (SupportedLanguage era)

mkTxInfoResult ∷ LedgerTxInfo era → TxInfoResult era

Construct PlutusTxInfo for all supported languages in this era.

lookupTxInfoResult ∷ EraPlutusTxInfo l era ⇒ SLanguage l → TxInfoResult era → Either (ContextError era) (PlutusTxInfo l)

TxInfo for the same language can be shared between executions of every script of the same version in a single transaction.

Note - The EraPlutusTxInfo is here only to enforce this function is not called with an unsupported plutus language version.

mkPlutusWithContext ∷ PlutusScript era → ScriptHash → PlutusPurpose AsIxItem era → LedgerTxInfo era → TxInfoResult era → (Data era, ExUnits) → CostModel → Either (ContextError era) PlutusWithContext

toPlutusWithContext ∷ ∀ l era. (EraPlutusTxInfo l era, AlonzoEraUTxO era) ⇒ Either (Plutus l) (PlutusRunnable l) → ScriptHash → PlutusPurpose AsIxItem era → LedgerTxInfo era → TxInfoResult era → (Data era, ExUnits) → CostModel → Either (ContextError era) PlutusWithContext

lookupTxInfoResultImpossible ∷ (HasCallStack, EraPlutusTxInfo l era) ⇒ SLanguage l → Either (ContextError era) (PlutusTxInfo l)

Helper function to use when implementing lookupTxInfoResult for plutus languages that are not supported by the era.

data SupportedLanguage era where

Constructors

SupportedLanguage ∷ EraPlutusTxInfo l era ⇒ SLanguage l → SupportedLanguage era

mkSupportedLanguageM ∷ ∀ era m. (EraPlutusContext era, MonadFail m) ⇒ Language → m (SupportedLanguage era)

supportedLanguages ∷ ∀ era. (HasCallStack, EraPlutusContext era) ⇒ NonEmpty (SupportedLanguage era)

mkSupportedPlutusScript ∷ ∀ l era. (HasCallStack, EraPlutusTxInfo l era) ⇒ Plutus l → PlutusScript era

This is just like mkPlutusScript, except it is guaranteed to be total through the enforcement of support by the type system and EraPlutusTxInfo type class instances for supported plutus versions.

mkSupportedBinaryPlutusScript ∷ ∀ era. (HasCallStack, AlonzoEraScript era) ⇒ SupportedLanguage era → PlutusBinary → PlutusScript era

This is just like mkBinaryPlutusScript, except it is guaranteed to be total through the enforcement of support by the type system and EraPlutusTxInfo type class instances (via calling `mkSupportedPlutusScript) for supported plutus versions.

Language dependent translation

type family PlutusTxInfo (l ∷ Language) where ...

Equations

PlutusTxInfo 'PlutusV1 = TxInfo

PlutusTxInfo 'PlutusV2 = TxInfo

PlutusTxInfo 'PlutusV3 = TxInfo

PlutusTxInfo 'PlutusV4 = TxInfo

type family PlutusTxCert (l ∷ Language) where ...

Equations

PlutusTxCert 'PlutusV1 = DCert

PlutusTxCert 'PlutusV2 = DCert

PlutusTxCert 'PlutusV3 = TxCert

PlutusTxCert 'PlutusV4 = TxCert

type family PlutusScriptPurpose (l ∷ Language) where ...

Equations

PlutusScriptPurpose 'PlutusV1 = ScriptPurpose

PlutusScriptPurpose 'PlutusV2 = ScriptPurpose

PlutusScriptPurpose 'PlutusV3 = ScriptPurpose

PlutusScriptPurpose 'PlutusV4 = ScriptPurpose

type family PlutusScriptContext (l ∷ Language) = (r ∷ Type) | r → l where ...

Equations

PlutusScriptContext 'PlutusV1 = ScriptContext

PlutusScriptContext 'PlutusV2 = ScriptContext

PlutusScriptContext 'PlutusV3 = ScriptContext

type family PlutusTxInInfo era (l ∷ Language) where ...

Equations

PlutusTxInInfo AlonzoEra PlutusV1 = Maybe TxInInfo

PlutusTxInInfo _ 'PlutusV1 = TxInInfo

PlutusTxInInfo _ 'PlutusV2 = TxInInfo

PlutusTxInInfo _ 'PlutusV3 = TxInInfo

PlutusTxInInfo _ 'PlutusV4 = TxInInfo