Cardano.Ledger.Alonzo.Plutus.Evaluate

Documentation

evalPlutusScripts ∷ EraTx era ⇒ Tx era → [PlutusWithContext (EraCrypto era)] → ScriptResult (EraCrypto era)

Evaluate a list of Plutus scripts. All scripts in the list must evaluate to True.

evalPlutusScriptsWithLogs ∷ EraTx era ⇒ Tx era → [PlutusWithContext (EraCrypto era)] → ([Text], ScriptResult (EraCrypto era))

data CollectError era

When collecting inputs for two phase scripts, 3 things can go wrong.

Constructors

NoRedeemer !(PlutusPurpose AsItem era)

NoWitness !(ScriptHash (EraCrypto era))

NoCostModel !Language

BadTranslation !(ContextError era)

collectPlutusScriptsWithContext ∷ ∀ era. (AlonzoEraTxBody era, AlonzoEraTxWits era, AlonzoEraUTxO era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, EraPlutusContext era) ⇒ EpochInfo (Either Text) → SystemStart → PParams era → Tx era → UTxO era → Either [CollectError era] [PlutusWithContext (EraCrypto era)]

lookupPlutusScript ∷ AlonzoEraScript era ⇒ Map (ScriptHash (EraCrypto era)) (Script era) → ScriptHash (EraCrypto era) → Maybe (PlutusScript era)

Given a script hash and a Map of available scripts, find the PlutusScript. Returns Nothing when script is missing or it is not a PlutusScript