Cardano.Ledger.Alonzo.Plutus.TxInfo

Documentation

data AlonzoContextError era

Constructors

TranslationLogicMissingInput !TxIn

TimeTranslationPastHorizon !Text

data TxOutSource

A transaction output can be translated because it is a newly created output, or because it is the output which is connected to a transaction input being spent.

Constructors

TxOutFromInput !TxIn

TxOutFromOutput !TxIx

transLookupTxOut ∷ ∀ era a. Inject (AlonzoContextError era) a ⇒ UTxO era → TxIn → Either a (TxOut era)

transTxOut ∷ (Value era ~ MaryValue, AlonzoEraTxOut era) ⇒ TxOut era → Maybe TxOut

Translate a TxOut. Returns Nothing if a Byron address is present in the TxOut.

transValidityInterval ∷ ∀ proxy era a. Inject (AlonzoContextError era) a ⇒ proxy era → EpochInfo (Either Text) → SystemStart → ValidityInterval → Either a POSIXTimeRange

Translate a validity interval to POSIX time

transPolicyID ∷ PolicyID → CurrencySymbol

transAssetName ∷ AssetName → TokenName

transMultiAsset ∷ MultiAsset → Value

transMintValue ∷ MultiAsset → Value

Hysterical raisins:

Previously transaction body contained a mint field with MaryValue instead of a MultiAsset, which has changed since then to just MultiAsset (because minting ADA makes no sense). However, if we don't preserve previous translation, scripts that previously succeeded will fail.

transValue ∷ MaryValue → Value

transWithdrawals ∷ Withdrawals → Map StakingCredential Integer

transDataPair ∷ (DataHash, Data era) → (DatumHash, Datum)

transTxCert ∷ (ShelleyEraTxCert era, AtMostEra "Babbage" era) ⇒ TxCert era → DCert

transTxCertCommon ∷ ShelleyEraTxCert era ⇒ TxCert era → Maybe DCert

Just like transTxCert, but do not translate certificates that were deprecated in Conway

transPlutusPurpose ∷ (EraPlutusTxInfo l era, PlutusTxCert l ~ DCert) ⇒ proxy l → ProtVer → AlonzoPlutusPurpose AsItem era → Either (ContextError era) ScriptPurpose

transTxBodyId ∷ EraTxBody era ⇒ TxBody era → TxId

transTxBodyCerts ∷ (EraPlutusTxInfo l era, EraTxBody era) ⇒ proxy l → ProtVer → TxBody era → Either (ContextError era) [PlutusTxCert l]

Translate all TxCerts from within a TxBody

transTxBodyWithdrawals ∷ EraTxBody era ⇒ TxBody era → [(StakingCredential, Integer)]

Translate all Withdrawals from within a TxBody

transTxBodyReqSignerHashes ∷ AlonzoEraTxBody era ⇒ TxBody era → [PubKeyHash]

Translate all required signers produced by reqSignerHashesTxBodyLs from within a TxBody

transTxWitsDatums ∷ AlonzoEraTxWits era ⇒ TxWits era → [(DatumHash, Datum)]

Translate all TxDatss from within TxWits

LgacyPlutusArgs helpers

toPlutusV1Args ∷ EraPlutusTxInfo 'PlutusV1 era ⇒ proxy 'PlutusV1 → ProtVer → TxInfo → PlutusPurpose AsIxItem era → Maybe (Data era) → Data era → Either (ContextError era) (PlutusArgs 'PlutusV1)

toLegacyPlutusArgs ∷ EraPlutusTxInfo l era ⇒ proxy l → ProtVer → (PlutusScriptPurpose l → PlutusScriptContext l) → PlutusPurpose AsIxItem era → Maybe (Data era) → Data era → Either (ContextError era) (LegacyPlutusArgs l)

Orphan instances

EraPlutusContext AlonzoEra

Instance details

Associated Types

type ContextError AlonzoEra = (r ∷ Type)

data TxInfoResult AlonzoEra

Methods

mkSupportedLanguage ∷ Language → Maybe (SupportedLanguage AlonzoEra)

mkTxInfoResult ∷ LedgerTxInfo AlonzoEra → TxInfoResult AlonzoEra

lookupTxInfoResult ∷ ∀ (l ∷ Language). EraPlutusTxInfo l AlonzoEra ⇒ SLanguage l → TxInfoResult AlonzoEra → Either (ContextError AlonzoEra) (PlutusTxInfo l)

mkPlutusWithContext ∷ PlutusScript AlonzoEra → ScriptHash → PlutusPurpose AsIxItem AlonzoEra → LedgerTxInfo AlonzoEra → TxInfoResult AlonzoEra → (Data AlonzoEra, ExUnits) → CostModel → Either (ContextError AlonzoEra) PlutusWithContext

EraPlutusTxInfo 'PlutusV1 AlonzoEra

Instance details

Methods

toPlutusTxCert ∷ proxy 'PlutusV1 → ProtVer → TxCert AlonzoEra → Either (ContextError AlonzoEra) (PlutusTxCert 'PlutusV1)

toPlutusScriptPurpose ∷ proxy 'PlutusV1 → ProtVer → PlutusPurpose AsIxItem AlonzoEra → Either (ContextError AlonzoEra) (PlutusScriptPurpose 'PlutusV1)

toPlutusTxInfo ∷ proxy 'PlutusV1 → LedgerTxInfo AlonzoEra → Either (ContextError AlonzoEra) (PlutusTxInfo 'PlutusV1)

toPlutusArgs ∷ proxy 'PlutusV1 → ProtVer → PlutusTxInfo 'PlutusV1 → PlutusPurpose AsIxItem AlonzoEra → Maybe (Data AlonzoEra) → Data AlonzoEra → Either (ContextError AlonzoEra) (PlutusArgs 'PlutusV1)

toPlutusTxInInfo ∷ proxy 'PlutusV1 → UTxO AlonzoEra → TxIn → Either (ContextError AlonzoEra) (PlutusTxInInfo AlonzoEra 'PlutusV1)