Cardano.Ledger.Alonzo.Rules

Documentation

data AlonzoBBODY era

data AlonzoBbodyPredFailure era

Constructors

ShelleyInAlonzoBbodyPredFailure (ShelleyBbodyPredFailure era)

TooManyExUnits (Mismatch 'RelLTEQ ExUnits)

newtype AlonzoBbodyEvent era

Constructors

ShelleyInAlonzoEvent (ShelleyBbodyEvent era)

alonzoBbodyTransition ∷ ∀ era. (STS (EraRule "BBODY" era), Signal (EraRule "BBODY" era) ~ Block BHeaderView era, InjectRuleFailure "BBODY" AlonzoBbodyPredFailure era, BaseM (EraRule "BBODY" era) ~ ShelleyBase, State (EraRule "BBODY" era) ~ ShelleyBbodyState era, Environment (EraRule "BBODY" era) ~ BbodyEnv era, Embed (EraRule "LEDGERS" era) (EraRule "BBODY" era), Environment (EraRule "LEDGERS" era) ~ ShelleyLedgersEnv era, State (EraRule "LEDGERS" era) ~ LedgerState era, Signal (EraRule "LEDGERS" era) ~ Seq (Tx era), EraBlockBody era, AlonzoEraTxWits era, BlockBody era ~ AlonzoBlockBody era, AlonzoEraPParams era) ⇒ TransitionRule (EraRule "BBODY" era)

data AlonzoLEDGER era

ledgerTransition ∷ ∀ (someLEDGER ∷ Type → Type) era. (STS (someLEDGER era), BaseM (someLEDGER era) ~ ShelleyBase, Signal (someLEDGER era) ~ Tx era, State (someLEDGER era) ~ LedgerState era, Environment (someLEDGER era) ~ LedgerEnv era, Embed (EraRule "UTXOW" era) (someLEDGER era), Embed (EraRule "DELEGS" era) (someLEDGER era), Environment (EraRule "DELEGS" era) ~ DelegsEnv era, State (EraRule "DELEGS" era) ~ CertState era, Signal (EraRule "DELEGS" era) ~ Seq (TxCert era), Environment (EraRule "UTXOW" era) ~ UtxoEnv era, State (EraRule "UTXOW" era) ~ UTxOState era, Signal (EraRule "UTXOW" era) ~ Tx era, AlonzoEraTx era) ⇒ TransitionRule (someLEDGER era)

An abstract Alonzo Era, Ledger transition. Fix someLedger at a concrete type to make it concrete.

data AlonzoUTXO era

data AlonzoUtxoPredFailure era

Constructors

BadInputsUTxO (Set TxIn)
The bad transaction inputs

OutsideValidityIntervalUTxO

Fields

ValidityInterval

transaction's validity interval

SlotNo

current slot

MaxTxSizeUTxO (Mismatch 'RelLTEQ Word32)

InputSetEmptyUTxO

FeeTooSmallUTxO (Mismatch 'RelGTEQ Coin)

ValueNotConservedUTxO (Mismatch 'RelEQ (Value era))

WrongNetwork
the set of addresses with incorrect network IDs

Fields

Network

the expected network id

(Set Addr)

the set of addresses with incorrect network IDs

WrongNetworkWithdrawal

Fields

Network

the expected network id

(Set RewardAccount)

the set of reward addresses with incorrect network IDs

OutputTooSmallUTxO [TxOut era]
list of supplied transaction outputs that are too small

UtxosFailure (PredicateFailure (EraRule "UTXOS" era))
Subtransition Failures

OutputBootAddrAttrsTooBig [TxOut era]
list of supplied bad transaction outputs

OutputTooBigUTxO [(Int, Int, TxOut era)]
list of supplied bad transaction output triples (actualSize,PParameterMaxValue,TxOut)

InsufficientCollateral

Fields

DeltaCoin

balance computed

Coin

the required collateral for the given fee

ScriptsNotPaidUTxO (UTxO era)
The UTxO entries which have the wrong kind of script

ExUnitsTooBigUTxO (Mismatch 'RelLTEQ ExUnits)

CollateralContainsNonADA (Value era)
The inputs marked for use as fees contain non-ADA tokens

WrongNetworkInTxBody (Mismatch 'RelEQ Network)
Wrong Network ID in body

OutsideForecast SlotNo
slot number outside consensus forecast range

TooManyCollateralInputs (Mismatch 'RelLTEQ Natural)
There are too many collateral inputs

NoCollateralInputs

allegraToAlonzoUtxoPredFailure ∷ ∀ t era. (EraRuleFailure "PPUP" era ~ t era, InjectRuleFailure "UTXOS" t era) ⇒ AllegraUtxoPredFailure era → AlonzoUtxoPredFailure era

newtype AlonzoUtxoEvent era

Constructors

UtxosEvent (Event (EraRule "UTXOS" era))

validateCollateralContainsNonADA ∷ (Foldable f, EraTxOut era) ⇒ f (TxOut era) → Test (AlonzoUtxoPredFailure era)

validateExUnitsTooBigUTxO ∷ (AlonzoEraTxWits era, EraTx era, AlonzoEraPParams era) ⇒ PParams era → Tx era → Test (AlonzoUtxoPredFailure era)

Ensure that execution units to not exceed the maximum allowed maxTxExUnits parameter.

totExunits tx ≤ maxTxExUnits pp

validateOutputTooBigUTxO ∷ (EraTxOut era, AlonzoEraPParams era, Foldable f) ⇒ PParams era → f (TxOut era) → Test (AlonzoUtxoPredFailure era)

Ensure that there are no TxOuts that have Value of size larger than MaxValSize. We use serialized length of Value because this Value size is being limited inside a serialized Tx.

∀ txout ∈ txouts txb, serSize (getValue txout) ≤ maxValSize pp

validateInsufficientCollateral ∷ (EraTxBody era, AlonzoEraPParams era) ⇒ PParams era → TxBody era → DeltaCoin → Test (AlonzoUtxoPredFailure era)

validateOutsideForecast

Arguments

∷ (MaryEraTxBody era, AlonzoEraTxWits era, EraTx era)

⇒ EpochInfo (Either a)

→ SlotNo
Current slot number

→ SystemStart

→ Tx era

→ Test (AlonzoUtxoPredFailure era)

If tx has non-native scripts, end of validity interval must translate to time

(_,i_f) := txvldt tx
◇ ∉ { txrdmrs tx, i_f } ⇒ epochInfoSlotToUTCTime epochInfo systemTime i_f ≠ ◇

validateScriptsNotPaidUTxO ∷ EraTxOut era ⇒ Map TxIn (TxOut era) → Test (AlonzoUtxoPredFailure era)

validateTooManyCollateralInputs ∷ AlonzoEraTxBody era ⇒ PParams era → TxBody era → Test (AlonzoUtxoPredFailure era)

Ensure that number of collaterals does not exceed the allowed maxCollInputs parameter.

‖collateral tx‖  ≤  maxCollInputs pp

validateWrongNetworkInTxBody ∷ AlonzoEraTxBody era ⇒ Network → TxBody era → Test (AlonzoUtxoPredFailure era)

Ensure if NetworkId is present in the txbody it matches the global NetworkId

(txnetworkid txb = NetworkId) ∨ (txnetworkid txb = ◇)

vKeyLocked ∷ EraTxOut era ⇒ TxOut era → Bool

data AlonzoUTXOS era

data AlonzoUtxosPredFailure era

Constructors

ValidationTagMismatch IsValid TagMismatchDescription
The isValid tag on the transaction is incorrect. The tag given here is that provided on the transaction (whereas evaluation of the scripts gives the opposite.). The Text tries to explain why it failed.

CollectErrors [CollectError era]
We could not find all the necessary inputs for a Plutus Script. Previous PredicateFailure tests should make this impossible, but the consequences of not detecting this means scripts get dropped, so things might validate that shouldn't. So we double check in the function collectTwoPhaseScriptInputs, it should find data for every Script.

UpdateFailure (EraRuleFailure "PPUP" era)

lbl2Phase ∷ Label

Indicates that this check depends only upon the signal to the transition, not the state or environment.

data TagMismatchDescription

Constructors

PassedUnexpectedly

FailedUnexpectedly (NonEmpty FailureDescription)

validBegin ∷ String

validEnd ∷ String

invalidBegin ∷ String

invalidEnd ∷ String

data AlonzoUtxosEvent era

Constructors

AlonzoPpupToUtxosEvent (EraRuleEvent "PPUP" era)

TotalDeposits (SafeHash EraIndependentTxBody) Coin

SuccessfulPlutusScriptsEvent (NonEmpty PlutusWithContext)

FailedPlutusScriptsEvent (NonEmpty PlutusWithContext)

TxUTxODiff
The UTxOs consumed and created by a signal tx

Fields

(UTxO era)

UTxO consumed

(UTxO era)

UTxO created

when2Phase ∷ Rule sts ctx () → Rule sts ctx ()

Construct a 2-phase predicate check.

Note that 2-phase predicate checks are by definition static.

data FailureDescription

Constructors

PlutusFailure Text ByteString

scriptFailureToFailureDescription ∷ ScriptFailure → FailureDescription

data AlonzoUTXOW era

newtype AlonzoUtxowEvent era

Constructors

WrappedShelleyEraEvent (ShelleyUtxowEvent era)

data AlonzoUtxowPredFailure era

The Predicate failure type in the Alonzo Era. It embeds the Predicate failure type of the Shelley Era, as they share some failure modes.

Constructors

ShelleyInAlonzoUtxowPredFailure (ShelleyUtxowPredFailure era)

MissingRedeemers [(PlutusPurpose AsItem era, ScriptHash)]
List of scripts for which no redeemers were supplied

MissingRequiredDatums

Fields

(Set DataHash)

Set of missing data hashes

(Set DataHash)

Set of received data hashes

NotAllowedSupplementalDatums

Fields

(Set DataHash)

Set of unallowed data hashes

(Set DataHash)

Set of acceptable supplemental data hashes

PPViewHashesDontMatch (Mismatch 'RelEQ (StrictMaybe ScriptIntegrityHash))

UnspendableUTxONoDatumHash (Set TxIn)
Set of transaction inputs that are TwoPhase scripts, and should have a DataHash but don't

ExtraRedeemers [PlutusPurpose AsIx era]
List of redeemers not needed

ScriptIntegrityHashMismatch (Mismatch 'RelEQ (StrictMaybe ScriptIntegrityHash)) (StrictMaybe ByteString)
The computed script integrity hash does not match the provided script integrity hash

hasExactSetOfRedeemers ∷ ∀ era. AlonzoEraTx era ⇒ Tx era → ScriptsProvided era → AlonzoScriptsNeeded era → Test (AlonzoUtxowPredFailure era)

missingRequiredDatums ∷ ∀ era. (AlonzoEraTx era, AlonzoEraUTxO era) ⇒ UTxO era → Tx era → Test (AlonzoUtxowPredFailure era)

checkScriptIntegrityHash ∷ ∀ era. AlonzoEraTx era ⇒ Tx era → PParams era → StrictMaybe (ScriptIntegrity era) → Test (AlonzoUtxowPredFailure era)