Cardano.Ledger.Alonzo.Rules.Utxo

Documentation

data AlonzoUTXO era

data AlonzoUtxoPredFailure era

Constructors

BadInputsUTxO !(Set (TxIn (EraCrypto era)))
The bad transaction inputs

OutsideValidityIntervalUTxO

Fields

!ValidityInterval

transaction's validity interval

!SlotNo

current slot

MaxTxSizeUTxO

Fields

!Integer

the actual transaction size

!Integer

the max transaction size

InputSetEmptyUTxO

FeeTooSmallUTxO

Fields

!Coin

the minimum fee for this transaction

!Coin

the fee supplied in this transaction

ValueNotConservedUTxO

Fields

!(Value era)

the Coin consumed by this transaction

!(Value era)

the Coin produced by this transaction

WrongNetwork
the set of addresses with incorrect network IDs

Fields

!Network

the expected network id

!(Set (Addr (EraCrypto era)))

the set of addresses with incorrect network IDs

WrongNetworkWithdrawal

Fields

!Network

the expected network id

!(Set (RewardAccount (EraCrypto era)))

the set of reward addresses with incorrect network IDs

OutputTooSmallUTxO ![TxOut era]
list of supplied transaction outputs that are too small

UtxosFailure (PredicateFailure (EraRule "UTXOS" era))
Subtransition Failures

OutputBootAddrAttrsTooBig ![TxOut era]
list of supplied bad transaction outputs

TriesToForgeADA

OutputTooBigUTxO ![(Integer, Integer, TxOut era)]
list of supplied bad transaction output triples (actualSize,PParameterMaxValue,TxOut)

InsufficientCollateral

Fields

!Coin

balance computed

!Coin

the required collateral for the given fee

ScriptsNotPaidUTxO !(UTxO era)
The UTxO entries which have the wrong kind of script

ExUnitsTooBigUTxO

Fields

!ExUnits

Max EXUnits from the protocol parameters

!ExUnits

EXUnits supplied

CollateralContainsNonADA !(Value era)
The inputs marked for use as fees contain non-ADA tokens

WrongNetworkInTxBody
Wrong Network ID in body

Fields

!Network

Actual Network ID

!Network

Network ID in transaction body

OutsideForecast !SlotNo
slot number outside consensus forecast range

TooManyCollateralInputs
There are too many collateral inputs

Fields

!Natural

Max allowed collateral inputs

!Natural

Number of collateral inputs

NoCollateralInputs

allegraToAlonzoUtxoPredFailure ∷ ∀ t era. (EraRuleFailure "PPUP" era ~ t era, InjectRuleFailure "UTXOS" t era) ⇒ AllegraUtxoPredFailure era → AlonzoUtxoPredFailure era

newtype AlonzoUtxoEvent era

Constructors

UtxosEvent (Event (EraRule "UTXOS" era))

validateCollateralContainsNonADA ∷ (Foldable f, EraTxOut era) ⇒ f (TxOut era) → Test (AlonzoUtxoPredFailure era)

validateExUnitsTooBigUTxO ∷ (AlonzoEraTxWits era, EraTx era, AlonzoEraPParams era) ⇒ PParams era → Tx era → Test (AlonzoUtxoPredFailure era)

Ensure that execution units to not exceed the maximum allowed maxTxExUnits parameter.

totExunits tx ≤ maxTxExUnits pp

validateOutputTooBigUTxO ∷ (EraTxOut era, AlonzoEraPParams era, Foldable f) ⇒ PParams era → f (TxOut era) → Test (AlonzoUtxoPredFailure era)

Ensure that there are no TxOuts that have Value of size larger than MaxValSize. We use serialized length of Value because this Value size is being limited inside a serialized Tx.

∀ txout ∈ txouts txb, serSize (getValue txout) ≤ maxValSize pp

validateInsufficientCollateral ∷ (EraTxBody era, AlonzoEraPParams era) ⇒ PParams era → TxBody era → Coin → Test (AlonzoUtxoPredFailure era)

validateOutsideForecast

Arguments

∷ (MaryEraTxBody era, AlonzoEraTxWits era, EraTx era)

⇒ EpochInfo (Either a)

→ SlotNo
Current slot number

→ SystemStart

→ Tx era

→ Test (AlonzoUtxoPredFailure era)

If tx has non-native scripts, end of validity interval must translate to time

(_,i_f) := txvldt tx
◇ ∉ { txrdmrs tx, i_f } ⇒ epochInfoSlotToUTCTime epochInfo systemTime i_f ≠ ◇

validateScriptsNotPaidUTxO ∷ EraTxOut era ⇒ Map (TxIn (EraCrypto era)) (TxOut era) → Test (AlonzoUtxoPredFailure era)

validateTooManyCollateralInputs ∷ AlonzoEraTxBody era ⇒ PParams era → TxBody era → Test (AlonzoUtxoPredFailure era)

Ensure that number of collaterals does not exceed the allowed maxCollInputs parameter.

‖collateral tx‖  ≤  maxCollInputs pp

validateWrongNetworkInTxBody ∷ AlonzoEraTxBody era ⇒ Network → TxBody era → Test (AlonzoUtxoPredFailure era)

Ensure if NetworkId is present in the txbody it matches the global NetworkId

(txnetworkid txb = NetworkId) ∨ (txnetworkid txb = ◇)

vKeyLocked ∷ EraTxOut era ⇒ TxOut era → Bool

Orphan instances

InjectRuleFailure "UTXO" AllegraUtxoPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ AllegraUtxoPredFailure (AlonzoEra c) → EraRuleFailure "UTXO" (AlonzoEra c)

InjectRuleFailure "UTXO" AlonzoUtxosPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxosPredFailure (AlonzoEra c) → EraRuleFailure "UTXO" (AlonzoEra c)

InjectRuleFailure "UTXO" ShelleyPpupPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ ShelleyPpupPredFailure (AlonzoEra c) → EraRuleFailure "UTXO" (AlonzoEra c)

InjectRuleFailure "UTXO" ShelleyUtxoPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxoPredFailure (AlonzoEra c) → EraRuleFailure "UTXO" (AlonzoEra c)

(EraUTxO era, AlonzoEraTx era, Embed (EraRule "UTXOS" era) (AlonzoUTXO era), Environment (EraRule "UTXOS" era) ~ UtxoEnv era, State (EraRule "UTXOS" era) ~ UTxOState era, Signal (EraRule "UTXOS" era) ~ Tx era, EraRule "UTXO" era ~ AlonzoUTXO era, InjectRuleFailure "UTXO" ShelleyUtxoPredFailure era, InjectRuleFailure "UTXO" AlonzoUtxoPredFailure era, InjectRuleFailure "UTXO" AllegraUtxoPredFailure era, ProtVerAtMost era 8) ⇒ STS (AlonzoUTXO era)

Instance details

Associated Types

type State (AlonzoUTXO era)

type Signal (AlonzoUTXO era)

type Environment (AlonzoUTXO era)

type BaseM (AlonzoUTXO era) ∷ Type → Type

type Event (AlonzoUTXO era)

type PredicateFailure (AlonzoUTXO era)

Methods

initialRules ∷ [InitialRule (AlonzoUTXO era)]

transitionRules ∷ [TransitionRule (AlonzoUTXO era)]

assertions ∷ [Assertion (AlonzoUTXO era)]

renderAssertionViolation ∷ AssertionViolation (AlonzoUTXO era) → String

(Era era, STS (AlonzoUTXOS era), PredicateFailure (EraRule "UTXOS" era) ~ AlonzoUtxosPredFailure era, Event (EraRule "UTXOS" era) ~ Event (AlonzoUTXOS era)) ⇒ Embed (AlonzoUTXOS era) (AlonzoUTXO era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (AlonzoUTXOS era) → PredicateFailure (AlonzoUTXO era)

wrapEvent ∷ Event (AlonzoUTXOS era) → Event (AlonzoUTXO era)