Cardano.Ledger.Alonzo.Rules.Utxos

Documentation

data AlonzoUTXOS era

data AlonzoUtxosPredFailure era

Constructors

ValidationTagMismatch IsValid TagMismatchDescription
The isValid tag on the transaction is incorrect. The tag given here is that provided on the transaction (whereas evaluation of the scripts gives the opposite.). The Text tries to explain why it failed.

CollectErrors [CollectError era]
We could not find all the necessary inputs for a Plutus Script. Previous PredicateFailure tests should make this impossible, but the consequences of not detecting this means scripts get dropped, so things might validate that shouldn't. So we double check in the function collectTwoPhaseScriptInputs, it should find data for every Script.

UpdateFailure (EraRuleFailure "PPUP" era)

lbl2Phase ∷ Label

Indicates that this check depends only upon the signal to the transition, not the state or environment.

data TagMismatchDescription

Constructors

PassedUnexpectedly

FailedUnexpectedly (NonEmpty FailureDescription)

validBegin ∷ String

validEnd ∷ String

invalidBegin ∷ String

invalidEnd ∷ String

data AlonzoUtxosEvent era

Constructors

AlonzoPpupToUtxosEvent (EraRuleEvent "PPUP" era)

TotalDeposits (SafeHash (EraCrypto era) EraIndependentTxBody) Coin

SuccessfulPlutusScriptsEvent (NonEmpty (PlutusWithContext (EraCrypto era)))

FailedPlutusScriptsEvent (NonEmpty (PlutusWithContext (EraCrypto era)))

TxUTxODiff
The UTxOs consumed and created by a signal tx

Fields

(UTxO era)

UTxO consumed

(UTxO era)

UTxO created

when2Phase ∷ Rule sts ctx () → Rule sts ctx ()

Construct a 2-phase predicate check.

Note that 2-phase predicate checks are by definition static.

data FailureDescription

Constructors

PlutusFailure Text ByteString

scriptFailureToFailureDescription ∷ Crypto c ⇒ ScriptFailure c → FailureDescription

Orphan instances

InjectRuleFailure "UTXOS" ShelleyPpupPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ ShelleyPpupPredFailure (AlonzoEra c) → EraRuleFailure "UTXOS" (AlonzoEra c)

(AlonzoEraTx era, AlonzoEraPParams era, ShelleyEraTxBody era, AlonzoEraUTxO era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, AlonzoEraScript era, TxCert era ~ ShelleyTxCert era, EraGov era, GovState era ~ ShelleyGovState era, State (EraRule "PPUP" era) ~ ShelleyGovState era, Embed (EraRule "PPUP" era) (AlonzoUTXOS era), Environment (EraRule "PPUP" era) ~ PpupEnv era, Signal (EraRule "PPUP" era) ~ StrictMaybe (Update era), EncCBOR (PredicateFailure (EraRule "PPUP" era)), Eq (EraRuleFailure "PPUP" era), Show (EraRuleFailure "PPUP" era), EraPlutusContext era) ⇒ STS (AlonzoUTXOS era)

Instance details

Associated Types

type State (AlonzoUTXOS era)

type Signal (AlonzoUTXOS era)

type Environment (AlonzoUTXOS era)

type BaseM (AlonzoUTXOS era) ∷ Type → Type

type Event (AlonzoUTXOS era)

type PredicateFailure (AlonzoUTXOS era)

Methods

initialRules ∷ [InitialRule (AlonzoUTXOS era)]

transitionRules ∷ [TransitionRule (AlonzoUTXOS era)]

assertions ∷ [Assertion (AlonzoUTXOS era)]

renderAssertionViolation ∷ AssertionViolation (AlonzoUTXOS era) → String

(Era era, STS (ShelleyPPUP era), EraRuleFailure "PPUP" era ~ ShelleyPpupPredFailure era, Event (EraRule "PPUP" era) ~ Event (ShelleyPPUP era), EraRuleEvent "PPUP" era ~ PpupEvent era) ⇒ Embed (ShelleyPPUP era) (AlonzoUTXOS era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyPPUP era) → PredicateFailure (AlonzoUTXOS era)

wrapEvent ∷ Event (ShelleyPPUP era) → Event (AlonzoUTXOS era)