Cardano.Ledger.Alonzo.Rules.Utxow

Documentation

data AlonzoUTXOW era

newtype AlonzoUtxowEvent era

Constructors

WrappedShelleyEraEvent (ShelleyUtxowEvent era)

data AlonzoUtxowPredFailure era

The Predicate failure type in the Alonzo Era. It embeds the Predicate failure type of the Shelley Era, as they share some failure modes.

Constructors

ShelleyInAlonzoUtxowPredFailure !(ShelleyUtxowPredFailure era)

MissingRedeemers ![(PlutusPurpose AsItem era, ScriptHash (EraCrypto era))]
List of scripts for which no redeemers were supplied

MissingRequiredDatums

Fields

!(Set (DataHash (EraCrypto era)))

Set of missing data hashes

!(Set (DataHash (EraCrypto era)))

Set of received data hashes

NotAllowedSupplementalDatums

Fields

!(Set (DataHash (EraCrypto era)))

Set of unallowed data hashes

!(Set (DataHash (EraCrypto era)))

Set of acceptable supplemental data hashes

PPViewHashesDontMatch

Fields

!(StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

The PPHash in the TxBody

!(StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

Computed from the current Protocol Parameters

MissingRequiredSigners (Set (KeyHash 'Witness (EraCrypto era)))
Set of witnesses which were needed and not supplied

UnspendableUTxONoDatumHash (Set (TxIn (EraCrypto era)))
Set of transaction inputs that are TwoPhase scripts, and should have a DataHash but don't

ExtraRedeemers ![PlutusPurpose AsIx era]
List of redeemers not needed

hasExactSetOfRedeemers ∷ ∀ era. AlonzoEraTx era ⇒ Tx era → ScriptsProvided era → AlonzoScriptsNeeded era → Test (AlonzoUtxowPredFailure era)

missingRequiredDatums ∷ ∀ era. (AlonzoEraTx era, AlonzoEraUTxO era) ⇒ UTxO era → Tx era → Test (AlonzoUtxowPredFailure era)

ppViewHashesMatch ∷ ∀ era. AlonzoEraTx era ⇒ Tx era → PParams era → ScriptsProvided era → Set (ScriptHash (EraCrypto era)) → Test (AlonzoUtxowPredFailure era)

requiredSignersAreWitnessed ∷ AlonzoEraTxBody era ⇒ TxBody era → Set (KeyHash 'Witness (EraCrypto era)) → Test (AlonzoUtxowPredFailure era)

Deprecated: As no longer used. validateNeededWitnesses now handles this check

Orphan instances

InjectRuleFailure "UTXOW" AllegraUtxoPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ AllegraUtxoPredFailure (AlonzoEra c) → EraRuleFailure "UTXOW" (AlonzoEra c)

InjectRuleFailure "UTXOW" AlonzoUtxoPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxoPredFailure (AlonzoEra c) → EraRuleFailure "UTXOW" (AlonzoEra c)

InjectRuleFailure "UTXOW" AlonzoUtxosPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxosPredFailure (AlonzoEra c) → EraRuleFailure "UTXOW" (AlonzoEra c)

InjectRuleFailure "UTXOW" ShelleyPpupPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ ShelleyPpupPredFailure (AlonzoEra c) → EraRuleFailure "UTXOW" (AlonzoEra c)

InjectRuleFailure "UTXOW" ShelleyUtxoPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxoPredFailure (AlonzoEra c) → EraRuleFailure "UTXOW" (AlonzoEra c)

InjectRuleFailure "UTXOW" ShelleyUtxowPredFailure (AlonzoEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxowPredFailure (AlonzoEra c) → EraRuleFailure "UTXOW" (AlonzoEra c)

(AlonzoEraTx era, EraTxAuxData era, AlonzoEraUTxO era, ShelleyEraTxBody era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, Signable (DSIGN (EraCrypto era)) (Hash (HASH (EraCrypto era)) EraIndependentTxBody), EraRule "UTXOW" era ~ AlonzoUTXOW era, InjectRuleFailure "UTXOW" ShelleyUtxowPredFailure era, InjectRuleFailure "UTXOW" AlonzoUtxowPredFailure era, Embed (EraRule "UTXO" era) (AlonzoUTXOW era), Environment (EraRule "UTXO" era) ~ UtxoEnv era, State (EraRule "UTXO" era) ~ UTxOState era, Signal (EraRule "UTXO" era) ~ Tx era) ⇒ STS (AlonzoUTXOW era)

Instance details

Associated Types

type State (AlonzoUTXOW era)

type Signal (AlonzoUTXOW era)

type Environment (AlonzoUTXOW era)

type BaseM (AlonzoUTXOW era) ∷ Type → Type

type Event (AlonzoUTXOW era)

type PredicateFailure (AlonzoUTXOW era)

Methods

initialRules ∷ [InitialRule (AlonzoUTXOW era)]

transitionRules ∷ [TransitionRule (AlonzoUTXOW era)]

assertions ∷ [Assertion (AlonzoUTXOW era)]

renderAssertionViolation ∷ AssertionViolation (AlonzoUTXOW era) → String

(Era era, STS (AlonzoUTXO era), PredicateFailure (EraRule "UTXO" era) ~ AlonzoUtxoPredFailure era, Event (EraRule "UTXO" era) ~ AlonzoUtxoEvent era, BaseM (AlonzoUTXOW era) ~ ShelleyBase, PredicateFailure (AlonzoUTXOW era) ~ AlonzoUtxowPredFailure era, Event (AlonzoUTXOW era) ~ AlonzoUtxowEvent era) ⇒ Embed (AlonzoUTXO era) (AlonzoUTXOW era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (AlonzoUTXO era) → PredicateFailure (AlonzoUTXOW era)

wrapEvent ∷ Event (AlonzoUTXO era) → Event (AlonzoUTXOW era)