Cardano.Ledger.Alonzo.Tx

Description

This module exports implementations of many of the functions outlined in the Alonzo specification. The link to source of the specification https://github.com/intersectmbo/cardano-ledger/tree/master/eras/alonzo/formal-spec The most recent version of the document can be found here: https://github.com/intersectmbo/cardano-ledger/releases/latest/download/alonzo-ledger.pdf The functions can be found in Figures in that document, and sections of this code refer to those figures.

Documentation

data CostModel

A language dependent cost model for the Plutus evaluator. Note that the EvaluationContext is entirely dependent on the cost model parameters (ie the Map Text Integer) and that this type uses the smart constructor mkCostModel to hide the evaluation context.

getLanguageView ∷ AlonzoEraPParams era ⇒ PParams era → Language → LangDepView

data Data era

type DataHash = SafeHash EraIndependentData

newtype IsValid

Tag indicating whether non-native scripts in this transaction are expected to validate. This is added by the block creator when constructing the block.

Constructors

IsValid Bool

hashData ∷ Data era → DataHash

nonNativeLanguages ∷ [Language]

hashScriptIntegrity ∷ ∀ era. AlonzoEraScript era ⇒ Set LangDepView → Redeemers era → TxDats era → StrictMaybe ScriptIntegrityHash

data EraIndependentScriptIntegrity

data ScriptIntegrity era

A ScriptIntegrityHash is the hash of three things. The first two come from the witnesses and the last comes from the Protocol Parameters.

Constructors

ScriptIntegrity !(Redeemers era) !(TxDats era) !(Set LangDepView)

type ScriptIntegrityHash = SafeHash EraIndependentScriptIntegrity

data AlonzoTx era

Constructors

AlonzoTx

Fields

body ∷ !(TxBody era)

wits ∷ !(TxWits era)

isValid ∷ !IsValid

auxiliaryData ∷ !(StrictMaybe (TxAuxData era))

class (EraTx era, AlonzoEraTxBody era, AlonzoEraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTx era where

Methods

isValidTxL ∷ Lens' (Tx era) IsValid

mkBasicAlonzoTx ∷ Monoid (TxWits era) ⇒ TxBody era → AlonzoTx era

bodyAlonzoTxL ∷ Lens' (AlonzoTx era) (TxBody era)

TxBody setter and getter for AlonzoTx.

witsAlonzoTxL ∷ Lens' (AlonzoTx era) (TxWits era)

TxWits setter and getter for AlonzoTx.

auxDataAlonzoTxL ∷ Lens' (AlonzoTx era) (StrictMaybe (TxAuxData era))

TxAuxData setter and getter for AlonzoTx.

sizeAlonzoTxF ∷ ∀ era. EraTx era ⇒ SimpleGetter (AlonzoTx era) Integer

txsize computes the length of the serialised bytes (for estimations)

wireSizeAlonzoTxF ∷ ∀ era. EraTx era ⇒ SimpleGetter (AlonzoTx era) Word32

txsize computes the length of the serialised bytes (actual size)

isValidAlonzoTxL ∷ Lens' (AlonzoTx era) IsValid

txdats' ∷ Era era ⇒ AlonzoTxWits era → TxDats era

txscripts' ∷ Era era ⇒ AlonzoTxWits era → Map ScriptHash (Script era)

txrdmrs ∷ AlonzoEraScript era ⇒ AlonzoTxWits era → Redeemers era

data AlonzoTxBody era where

Bundled Patterns

pattern AlonzoTxBody ∷ ∀ era. (EraTxOut era, EraTxCert era) ⇒ Set TxIn → Set TxIn → StrictSeq (TxOut era) → StrictSeq (TxCert era) → Withdrawals → Coin → ValidityInterval → StrictMaybe (Update era) → Set (KeyHash 'Witness) → MultiAsset → StrictMaybe ScriptIntegrityHash → StrictMaybe TxAuxDataHash → StrictMaybe Network → AlonzoTxBody era

totExUnits ∷ (EraTx era, AlonzoEraTxWits era) ⇒ Tx era → ExUnits

alonzoMinFeeTx ∷ (EraTx era, AlonzoEraTxWits era, AlonzoEraPParams era) ⇒ PParams era → Tx era → Coin

txouts ∷ EraTxBody era ⇒ TxBody era → UTxO era

Compute the transaction outputs of a transaction.

alonzoSegwitTx ∷ AlonzoEraTx era ⇒ Annotator (TxBody era) → Annotator (TxWits era) → IsValid → Maybe (Annotator (TxAuxData era)) → Annotator (Tx era)

Construct an annotated Alonzo style transaction.

toCBORForSizeComputation ∷ (EncCBOR (TxBody era), EncCBOR (TxWits era), EncCBOR (TxAuxData era)) ⇒ AlonzoTx era → Encoding

This ensures that the size of transactions from Mary is unchanged. The individual components all store their bytes; the only work we do in this function is concatenating

toCBORForMempoolSubmission ∷ (EncCBOR (TxBody era), EncCBOR (TxWits era), EncCBOR (TxAuxData era)) ⇒ AlonzoTx era → Encoding

Encode to CBOR for the purposes of transmission from node to node, or from wallet to node.

Note that this serialisation is neither the serialisation used on-chain (where Txs are deconstructed using segwit), nor the serialisation used for computing the transaction size (which omits the IsValid field for compatibility with Mary - see toCBORForSizeComputation).

alonzoEqTxRaw ∷ AlonzoEraTx era ⇒ Tx era → Tx era → Bool

Orphan instances

EraTx AlonzoEra

Instance details

Associated Types

type Tx AlonzoEra = (r ∷ Type)

type TxUpgradeError AlonzoEra

Methods

mkBasicTx ∷ TxBody AlonzoEra → Tx AlonzoEra

bodyTxL ∷ Lens' (Tx AlonzoEra) (TxBody AlonzoEra)

witsTxL ∷ Lens' (Tx AlonzoEra) (TxWits AlonzoEra)

auxDataTxL ∷ Lens' (Tx AlonzoEra) (StrictMaybe (TxAuxData AlonzoEra))

sizeTxF ∷ SimpleGetter (Tx AlonzoEra) Integer

wireSizeTxF ∷ SimpleGetter (Tx AlonzoEra) Word32

validateNativeScript ∷ Tx AlonzoEra → NativeScript AlonzoEra → Bool

getMinFeeTx ∷ PParams AlonzoEra → Tx AlonzoEra → Int → Coin

upgradeTx ∷ Tx (PreviousEra AlonzoEra) → Either (TxUpgradeError AlonzoEra) (Tx AlonzoEra)