Cardano.Ledger.Alonzo.TxAuxData

AlonzoTxAuxData

data AlonzoTxAuxData era where

Bundled Patterns

pattern AlonzoTxAuxData ∷ ∀ era. (HasCallStack, AlonzoEraScript era) ⇒ Map Word64 Metadatum → StrictSeq (Timelock era) → Map Language (NonEmpty PlutusBinary) → AlonzoTxAuxData era
Construct auxiliary data. Make sure not to supply plutus script versions that are not supported in this era, because it will result in a runtime exception. Use mkAlonzoTxAuxData instead if you need runtime safety guarantees.

pattern AlonzoTxAuxData' ∷ ∀ era. Map Word64 Metadatum → StrictSeq (Timelock era) → Map Language (NonEmpty PlutusBinary) → AlonzoTxAuxData era

class AllegraEraTxAuxData era ⇒ AlonzoEraTxAuxData era where

Methods

plutusScriptsTxAuxDataL ∷ Lens' (TxAuxData era) (Map Language (NonEmpty PlutusBinary))

data AlonzoTxAuxDataRaw era

mkAlonzoTxAuxData ∷ ∀ f era. (Foldable f, AlonzoEraScript era) ⇒ Map Word64 Metadatum → f (AlonzoScript era) → AlonzoTxAuxData era

Helper function that will construct Auxiliary data from Metadatum map and a list of scripts.

Note that the relative order of same type scripts will be preserved.

hashAlonzoTxAuxData ∷ HashAnnotated x EraIndependentTxAuxData ⇒ x → TxAuxDataHash

Deprecated: In favor of hashTxAuxData

validateAlonzoTxAuxData ∷ (AlonzoEraScript era, Script era ~ AlonzoScript era) ⇒ ProtVer → AlonzoTxAuxData era → Bool

getAlonzoTxAuxDataScripts ∷ ∀ era. AlonzoEraScript era ⇒ AlonzoTxAuxData era → StrictSeq (AlonzoScript era)

translateAlonzoTxAuxData ∷ (AlonzoEraScript era1, AlonzoEraScript era2) ⇒ AlonzoTxAuxData era1 → AlonzoTxAuxData era2

metadataAlonzoTxAuxDataL ∷ ∀ era. Era era ⇒ Lens' (AlonzoTxAuxData era) (Map Word64 Metadatum)

timelockScriptsAlonzoTxAuxDataL ∷ ∀ era. Era era ⇒ Lens' (AlonzoTxAuxData era) (StrictSeq (Timelock era))

plutusScriptsAllegraTxAuxDataL ∷ ∀ era. Era era ⇒ Lens' (AlonzoTxAuxData era) (Map Language (NonEmpty PlutusBinary))

Orphan instances

AllegraEraTxAuxData AlonzoEra

Instance details

Methods

timelockScriptsTxAuxDataL ∷ Lens' (TxAuxData AlonzoEra) (StrictSeq (Timelock AlonzoEra))

EraTxAuxData AlonzoEra

Instance details

Associated Types

type TxAuxData AlonzoEra = (r ∷ Type)

Methods

mkBasicTxAuxData ∷ TxAuxData AlonzoEra

metadataTxAuxDataL ∷ Lens' (TxAuxData AlonzoEra) (Map Word64 Metadatum)

upgradeTxAuxData ∷ TxAuxData (PreviousEra AlonzoEra) → TxAuxData AlonzoEra

validateTxAuxData ∷ ProtVer → TxAuxData AlonzoEra → Bool