Cardano.Ledger.Alonzo.TxBody

Documentation

data AlonzoTxOut era

Constructors

TxOutCompact' !(CompactAddr (EraCrypto era)) !(CompactForm (Value era))

TxOutCompactDH' !(CompactAddr (EraCrypto era)) !(CompactForm (Value era)) !(DataHash (EraCrypto era))

TxOut_AddrHash28_AdaOnly !(Credential 'Staking (EraCrypto era)) !Addr28Extra !(CompactForm Coin)

TxOut_AddrHash28_AdaOnly_DataHash32 !(Credential 'Staking (EraCrypto era)) !Addr28Extra !(CompactForm Coin) !DataHash32

Bundled Patterns

pattern TxOutCompact ∷ (Era era, Val (Value era), HasCallStack) ⇒ CompactAddr (EraCrypto era) → CompactForm (Value era) → AlonzoTxOut era

pattern AlonzoTxOut ∷ ∀ era. (Era era, Val (Value era), HasCallStack) ⇒ Addr (EraCrypto era) → Value era → StrictMaybe (DataHash (EraCrypto era)) → AlonzoTxOut era

pattern TxOutCompactDH ∷ (Era era, Val (Value era), HasCallStack) ⇒ CompactAddr (EraCrypto era) → CompactForm (Value era) → DataHash (EraCrypto era) → AlonzoTxOut era

class (AlonzoEraPParams era, EraTxOut era) ⇒ AlonzoEraTxOut era where

Methods

dataHashTxOutL ∷ Lens' (TxOut era) (StrictMaybe (DataHash (EraCrypto era)))

datumTxOutF ∷ SimpleGetter (TxOut era) (Datum era)

data Addr28Extra

data DataHash32

data AlonzoTxBody era where

Bundled Patterns

pattern AlonzoTxBody ∷ (EraTxOut era, EraTxCert era) ⇒ Set (TxIn (EraCrypto era)) → Set (TxIn (EraCrypto era)) → StrictSeq (TxOut era) → StrictSeq (TxCert era) → Withdrawals (EraCrypto era) → Coin → ValidityInterval → StrictMaybe (Update era) → Set (KeyHash 'Witness (EraCrypto era)) → MultiAsset (EraCrypto era) → StrictMaybe (ScriptIntegrityHash (EraCrypto era)) → StrictMaybe (AuxiliaryDataHash (EraCrypto era)) → StrictMaybe Network → AlonzoTxBody era

data AlonzoTxBodyRaw era

data AlonzoTxBodyUpgradeError

Constructors

ATBUEMinUTxOUpdated
The TxBody contains a protocol parameter update that attempts to update the min UTxO. Since this doesn't exist in Alonzo, we fail if an attempt is made to update it.

class (MaryEraTxBody era, AlonzoEraTxOut era) ⇒ AlonzoEraTxBody era where

Methods

collateralInputsTxBodyL ∷ Lens' (TxBody era) (Set (TxIn (EraCrypto era)))

reqSignerHashesTxBodyL ∷ Lens' (TxBody era) (Set (KeyHash 'Witness (EraCrypto era)))

scriptIntegrityHashTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

networkIdTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Network)

redeemerPointer ∷ TxBody era → PlutusPurpose AsItem era → StrictMaybe (PlutusPurpose AsIx era)

Deprecated: As no longer needed

This function is called rdptr in the spec. Given a TxBody and a plutus purpose with an item, we should be able to find the plutus purpose as in index

redeemerPointerInverse ∷ TxBody era → PlutusPurpose AsIx era → StrictMaybe (PlutusPurpose AsIxItem era)

This is an inverse of redeemerPointer. Given purpose as an index return it as an item.

class (ShelleyEraTxCert era, EraTxBody era, ProtVerAtMost era 8) ⇒ ShelleyEraTxBody era where

Methods

ttlTxBodyL ∷ Lens' (TxBody era) SlotNo

updateTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (Update era))

class EraTxBody era ⇒ AllegraEraTxBody era where

Methods

vldtTxBodyL ∷ Lens' (TxBody era) ValidityInterval

class AllegraEraTxBody era ⇒ MaryEraTxBody era where

Methods

mintTxBodyL ∷ Lens' (TxBody era) (MultiAsset (EraCrypto era))

mintValueTxBodyF ∷ SimpleGetter (TxBody era) (Value era)

mintedTxBodyF ∷ SimpleGetter (TxBody era) (Set (PolicyID (EraCrypto era)))

class Indexable elem container where

Methods

indexOf ∷ AsItem Word32 elem → container → StrictMaybe (AsIx Word32 elem)

fromIndex ∷ AsIx Word32 elem → container → StrictMaybe (AsIxItem Word32 elem)

inputs' ∷ AlonzoTxBody era → Set (TxIn (EraCrypto era))

collateral' ∷ AlonzoTxBody era → Set (TxIn (EraCrypto era))

outputs' ∷ AlonzoTxBody era → StrictSeq (TxOut era)

certs' ∷ AlonzoTxBody era → StrictSeq (TxCert era)

withdrawals' ∷ AlonzoTxBody era → Withdrawals (EraCrypto era)

txfee' ∷ AlonzoTxBody era → Coin

vldt' ∷ AlonzoTxBody era → ValidityInterval

update' ∷ AlonzoTxBody era → StrictMaybe (Update era)

reqSignerHashes' ∷ AlonzoTxBody era → Set (KeyHash 'Witness (EraCrypto era))

mint' ∷ AlonzoTxBody era → MultiAsset (EraCrypto era)

scriptIntegrityHash' ∷ AlonzoTxBody era → StrictMaybe (ScriptIntegrityHash (EraCrypto era))

adHash' ∷ AlonzoTxBody era → StrictMaybe (AuxiliaryDataHash (EraCrypto era))

txnetworkid' ∷ AlonzoTxBody era → StrictMaybe Network

getAdaOnly ∷ ∀ era. Val (Value era) ⇒ Proxy era → Value era → Maybe (CompactForm Coin)

decodeDataHash32 ∷ ∀ c. HashAlgorithm (HASH c) ⇒ DataHash32 → Maybe (DataHash c)

encodeDataHash32 ∷ ∀ c. HashAlgorithm (HASH c) ⇒ DataHash c → Maybe (SizeHash (HASH c) :~: 32, DataHash32)

encodeAddress28 ∷ ∀ c. HashAlgorithm (ADDRHASH c) ⇒ Network → PaymentCredential c → Maybe (SizeHash (ADDRHASH c) :~: 28, Addr28Extra)

decodeAddress28 ∷ ∀ c. HashAlgorithm (ADDRHASH c) ⇒ Credential 'Staking c → Addr28Extra → Maybe (Addr c)

viewCompactTxOut ∷ (Era era, Val (Value era)) ⇒ AlonzoTxOut era → (CompactAddr (EraCrypto era), CompactForm (Value era), StrictMaybe (DataHash (EraCrypto era)))

viewTxOut ∷ (Era era, Val (Value era)) ⇒ AlonzoTxOut era → (Addr (EraCrypto era), Value era, StrictMaybe (DataHash (EraCrypto era)))

data EraIndependentScriptIntegrity

type ScriptIntegrityHash c = SafeHash c EraIndependentScriptIntegrity

getAlonzoTxOutEitherAddr ∷ HashAlgorithm (ADDRHASH (EraCrypto era)) ⇒ AlonzoTxOut era → Either (Addr (EraCrypto era)) (CompactAddr (EraCrypto era))

utxoEntrySize ∷ AlonzoEraTxOut era ⇒ TxOut era → Integer

Compute an estimate of the size of storing one UTxO entry. This function implements the UTxO entry size estimate done by scaledMinDeposit in the ShelleyMA era

alonzoRedeemerPointer ∷ ∀ era. MaryEraTxBody era ⇒ TxBody era → AlonzoPlutusPurpose AsItem era → StrictMaybe (AlonzoPlutusPurpose AsIx era)

alonzoRedeemerPointerInverse ∷ MaryEraTxBody era ⇒ TxBody era → AlonzoPlutusPurpose AsIx era → StrictMaybe (AlonzoPlutusPurpose AsIxItem era)

Orphan instances

Crypto c ⇒ AllegraEraTxBody (AlonzoEra c)

Instance details

Methods

vldtTxBodyL ∷ Lens' (TxBody (AlonzoEra c)) ValidityInterval

Crypto c ⇒ EraTxBody (AlonzoEra c)

Instance details

Associated Types

type TxBody (AlonzoEra c) = (r ∷ Type)

type TxBodyUpgradeError (AlonzoEra c)

Methods

mkBasicTxBody ∷ TxBody (AlonzoEra c)

inputsTxBodyL ∷ Lens' (TxBody (AlonzoEra c)) (Set (TxIn (EraCrypto (AlonzoEra c))))

outputsTxBodyL ∷ Lens' (TxBody (AlonzoEra c)) (StrictSeq (TxOut (AlonzoEra c)))

feeTxBodyL ∷ Lens' (TxBody (AlonzoEra c)) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody (AlonzoEra c)) (Withdrawals (EraCrypto (AlonzoEra c)))

auxDataHashTxBodyL ∷ Lens' (TxBody (AlonzoEra c)) (StrictMaybe (AuxiliaryDataHash (EraCrypto (AlonzoEra c))))

spendableInputsTxBodyF ∷ SimpleGetter (TxBody (AlonzoEra c)) (Set (TxIn (EraCrypto (AlonzoEra c))))

allInputsTxBodyF ∷ SimpleGetter (TxBody (AlonzoEra c)) (Set (TxIn (EraCrypto (AlonzoEra c))))

certsTxBodyL ∷ Lens' (TxBody (AlonzoEra c)) (StrictSeq (TxCert (AlonzoEra c)))

getTotalDepositsTxBody ∷ PParams (AlonzoEra c) → (KeyHash 'StakePool (EraCrypto (AlonzoEra c)) → Bool) → TxBody (AlonzoEra c) → Coin

getTotalRefundsTxBody ∷ PParams (AlonzoEra c) → (Credential 'Staking (EraCrypto (AlonzoEra c)) → Maybe Coin) → (Credential 'DRepRole (EraCrypto (AlonzoEra c)) → Maybe Coin) → TxBody (AlonzoEra c) → Coin

getGenesisKeyHashCountTxBody ∷ TxBody (AlonzoEra c) → Int

upgradeTxBody ∷ TxBody (PreviousEra (AlonzoEra c)) → Either (TxBodyUpgradeError (AlonzoEra c)) (TxBody (AlonzoEra c))

Crypto c ⇒ MaryEraTxBody (AlonzoEra c)

Instance details

Methods

mintTxBodyL ∷ Lens' (TxBody (AlonzoEra c)) (MultiAsset (EraCrypto (AlonzoEra c)))

mintValueTxBodyF ∷ SimpleGetter (TxBody (AlonzoEra c)) (Value (AlonzoEra c))

mintedTxBodyF ∷ SimpleGetter (TxBody (AlonzoEra c)) (Set (PolicyID (EraCrypto (AlonzoEra c))))

Crypto c ⇒ ShelleyEraTxBody (AlonzoEra c)

Instance details

Methods

ttlTxBodyL ∷ Lens' (TxBody (AlonzoEra c)) SlotNo

updateTxBodyL ∷ Lens' (TxBody (AlonzoEra c)) (StrictMaybe (Update (AlonzoEra c)))