Cardano.Ledger.Alonzo.TxBody.Internal

Description

Provides Alonzo TxBody internals

Warning

This module is considered internal.

The contents of this module may change in any way whatsoever and without any warning between minor versions of this package.

Documentation

data AlonzoTxOut era

Constructors

TxOutCompact' !CompactAddr !(CompactForm (Value era))

TxOutCompactDH' !CompactAddr !(CompactForm (Value era)) !DataHash

TxOut_AddrHash28_AdaOnly !(Credential 'Staking) !Addr28Extra !(CompactForm Coin)

TxOut_AddrHash28_AdaOnly_DataHash32 !(Credential 'Staking) !Addr28Extra !(CompactForm Coin) !DataHash32

Bundled Patterns

pattern TxOutCompact ∷ (Era era, Val (Value era), HasCallStack) ⇒ CompactAddr → CompactForm (Value era) → AlonzoTxOut era

pattern AlonzoTxOut ∷ ∀ era. (Era era, Val (Value era), HasCallStack) ⇒ Addr → Value era → StrictMaybe DataHash → AlonzoTxOut era

pattern TxOutCompactDH ∷ (Era era, Val (Value era), HasCallStack) ⇒ CompactAddr → CompactForm (Value era) → DataHash → AlonzoTxOut era

class (AlonzoEraPParams era, EraTxOut era) ⇒ AlonzoEraTxOut era where

Methods

dataHashTxOutL ∷ Lens' (TxOut era) (StrictMaybe DataHash)

datumTxOutF ∷ SimpleGetter (TxOut era) (Datum era)

data Addr28Extra

data DataHash32

newtype AlonzoTxBody era

Constructors

TxBodyConstr (MemoBytes (AlonzoTxBodyRaw era))

Bundled Patterns

pattern AlonzoTxBody ∷ ∀ era. (EraTxOut era, EraTxCert era) ⇒ Set TxIn → Set TxIn → StrictSeq (TxOut era) → StrictSeq (TxCert era) → Withdrawals → Coin → ValidityInterval → StrictMaybe (Update era) → Set (KeyHash 'Witness) → MultiAsset → StrictMaybe ScriptIntegrityHash → StrictMaybe TxAuxDataHash → StrictMaybe Network → AlonzoTxBody era

data AlonzoTxBodyRaw era

Constructors

AlonzoTxBodyRaw

Fields

atbrInputs ∷ !(Set TxIn)

atbrCollateral ∷ !(Set TxIn)

atbrOutputs ∷ !(StrictSeq (TxOut era))

atbrCerts ∷ !(StrictSeq (TxCert era))

atbrWithdrawals ∷ !Withdrawals

atbrTxFee ∷ !Coin

atbrValidityInterval ∷ !ValidityInterval

atbrUpdate ∷ !(StrictMaybe (Update era))

atbrReqSignerHashes ∷ Set (KeyHash 'Witness)

atbrMint ∷ !MultiAsset

atbrScriptIntegrityHash ∷ !(StrictMaybe ScriptIntegrityHash)

atbrAuxDataHash ∷ !(StrictMaybe TxAuxDataHash)

atbrTxNetworkId ∷ !(StrictMaybe Network)

data AlonzoTxBodyUpgradeError

Constructors

ATBUEMinUTxOUpdated
The TxBody contains a protocol parameter update that attempts to update the min UTxO. Since this doesn't exist in Alonzo, we fail if an attempt is made to update it.

class (MaryEraTxBody era, AlonzoEraTxOut era) ⇒ AlonzoEraTxBody era where

Methods

collateralInputsTxBodyL ∷ Lens' (TxBody era) (Set TxIn)

reqSignerHashesTxBodyL ∷ Lens' (TxBody era) (Set (KeyHash 'Witness))

scriptIntegrityHashTxBodyL ∷ Lens' (TxBody era) (StrictMaybe ScriptIntegrityHash)

networkIdTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Network)

redeemerPointer ∷ TxBody era → PlutusPurpose AsItem era → StrictMaybe (PlutusPurpose AsIx era)

This function is called rdptr in the spec. Given a TxBody and a plutus purpose with an item, we should be able to find the plutus purpose as in index

redeemerPointerInverse ∷ TxBody era → PlutusPurpose AsIx era → StrictMaybe (PlutusPurpose AsIxItem era)

This is an inverse of redeemerPointer. Given purpose as an index return it as an item.

class (ShelleyEraTxCert era, EraTxBody era, ProtVerAtMost era 8) ⇒ ShelleyEraTxBody era where

Methods

ttlTxBodyL ∷ Lens' (TxBody era) SlotNo

updateTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (Update era))

class EraTxBody era ⇒ AllegraEraTxBody era where

Methods

vldtTxBodyL ∷ Lens' (TxBody era) ValidityInterval

class AllegraEraTxBody era ⇒ MaryEraTxBody era where

Methods

mintTxBodyL ∷ Lens' (TxBody era) MultiAsset

mintValueTxBodyF ∷ SimpleGetter (TxBody era) (Value era)

mintedTxBodyF ∷ SimpleGetter (TxBody era) (Set PolicyID)

class Indexable elem container where

Methods

indexOf ∷ AsItem Word32 elem → container → StrictMaybe (AsIx Word32 elem)

fromIndex ∷ AsIx Word32 elem → container → StrictMaybe (AsIxItem Word32 elem)

inputs' ∷ AlonzoTxBody era → Set TxIn

collateral' ∷ AlonzoTxBody era → Set TxIn

outputs' ∷ AlonzoTxBody era → StrictSeq (TxOut era)

certs' ∷ AlonzoTxBody era → StrictSeq (TxCert era)

withdrawals' ∷ AlonzoTxBody era → Withdrawals

txfee' ∷ AlonzoTxBody era → Coin

vldt' ∷ AlonzoTxBody era → ValidityInterval

update' ∷ AlonzoTxBody era → StrictMaybe (Update era)

reqSignerHashes' ∷ AlonzoTxBody era → Set (KeyHash 'Witness)

mint' ∷ AlonzoTxBody era → MultiAsset

scriptIntegrityHash' ∷ AlonzoTxBody era → StrictMaybe ScriptIntegrityHash

adHash' ∷ AlonzoTxBody era → StrictMaybe TxAuxDataHash

txnetworkid' ∷ AlonzoTxBody era → StrictMaybe Network

getAdaOnly ∷ ∀ era. Val (Value era) ⇒ Proxy era → Value era → Maybe (CompactForm Coin)

decodeDataHash32 ∷ DataHash32 → DataHash

encodeDataHash32 ∷ DataHash → DataHash32

encodeAddress28 ∷ Network → PaymentCredential → Addr28Extra

decodeAddress28 ∷ Credential 'Staking → Addr28Extra → Addr

viewCompactTxOut ∷ Val (Value era) ⇒ AlonzoTxOut era → (CompactAddr, CompactForm (Value era), StrictMaybe DataHash)

viewTxOut ∷ Val (Value era) ⇒ AlonzoTxOut era → (Addr, Value era, StrictMaybe DataHash)

data EraIndependentScriptIntegrity

type ScriptIntegrityHash = SafeHash EraIndependentScriptIntegrity

getAlonzoTxOutEitherAddr ∷ AlonzoTxOut era → Either Addr CompactAddr

utxoEntrySize ∷ AlonzoEraTxOut era ⇒ TxOut era → Integer

Compute an estimate of the size of storing one UTxO entry. This function implements the UTxO entry size estimate done by scaledMinDeposit in the ShelleyMA era

alonzoRedeemerPointer ∷ ∀ era. MaryEraTxBody era ⇒ TxBody era → AlonzoPlutusPurpose AsItem era → StrictMaybe (AlonzoPlutusPurpose AsIx era)

alonzoRedeemerPointerInverse ∷ MaryEraTxBody era ⇒ TxBody era → AlonzoPlutusPurpose AsIx era → StrictMaybe (AlonzoPlutusPurpose AsIxItem era)

Orphan instances

AllegraEraTxBody AlonzoEra

Instance details

Methods

vldtTxBodyL ∷ Lens' (TxBody AlonzoEra) ValidityInterval

EraTxBody AlonzoEra

Instance details

Associated Types

type TxBody AlonzoEra = (r ∷ Type)

type TxBodyUpgradeError AlonzoEra

Methods

mkBasicTxBody ∷ TxBody AlonzoEra

inputsTxBodyL ∷ Lens' (TxBody AlonzoEra) (Set TxIn)

outputsTxBodyL ∷ Lens' (TxBody AlonzoEra) (StrictSeq (TxOut AlonzoEra))

feeTxBodyL ∷ Lens' (TxBody AlonzoEra) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody AlonzoEra) Withdrawals

auxDataHashTxBodyL ∷ Lens' (TxBody AlonzoEra) (StrictMaybe TxAuxDataHash)

spendableInputsTxBodyF ∷ SimpleGetter (TxBody AlonzoEra) (Set TxIn)

allInputsTxBodyF ∷ SimpleGetter (TxBody AlonzoEra) (Set TxIn)

certsTxBodyL ∷ Lens' (TxBody AlonzoEra) (StrictSeq (TxCert AlonzoEra))

getTotalDepositsTxBody ∷ PParams AlonzoEra → (KeyHash 'StakePool → Bool) → TxBody AlonzoEra → Coin

getTotalRefundsTxBody ∷ PParams AlonzoEra → (Credential 'Staking → Maybe Coin) → (Credential 'DRepRole → Maybe Coin) → TxBody AlonzoEra → Coin

getGenesisKeyHashCountTxBody ∷ TxBody AlonzoEra → Int

upgradeTxBody ∷ TxBody (PreviousEra AlonzoEra) → Either (TxBodyUpgradeError AlonzoEra) (TxBody AlonzoEra)

MaryEraTxBody AlonzoEra

Instance details

Methods

mintTxBodyL ∷ Lens' (TxBody AlonzoEra) MultiAsset

mintValueTxBodyF ∷ SimpleGetter (TxBody AlonzoEra) (Value AlonzoEra)

mintedTxBodyF ∷ SimpleGetter (TxBody AlonzoEra) (Set PolicyID)

ShelleyEraTxBody AlonzoEra

Instance details

Methods

ttlTxBodyL ∷ Lens' (TxBody AlonzoEra) SlotNo

updateTxBodyL ∷ Lens' (TxBody AlonzoEra) (StrictMaybe (Update AlonzoEra))