Cardano.Ledger.Alonzo.TxInfo

Documentation

data TxOutSource crypto

A transaction output can be translated because it is a newly created output, or because it is the output which is connected to a transaction input being spent.

Constructors

TxOutFromInput !(TxIn crypto)

TxOutFromOutput !TxIx

data TranslationError crypto

Constructors

ByronTxOutInContext !(TxOutSource crypto)

TranslationLogicMissingInput !(TxIn crypto)

RdmrPtrPointsToNothing !RdmrPtr

LanguageNotSupported !Language

InlineDatumsNotSupported !(TxOutSource crypto)

ReferenceScriptsNotSupported !(TxOutSource crypto)

ReferenceInputsNotSupported !(Set (TxIn crypto))

TimeTranslationPastHorizon !Text

transDataHash :: StrictMaybe (DataHash c) -> Maybe DatumHash

transDataHash' :: DataHash c -> DatumHash

transKeyHash :: KeyHash d c -> PubKeyHash

transScriptHash :: ScriptHash c -> ValidatorHash

transSafeHash :: SafeHash c i -> BuiltinByteString

transHash :: Hash h a -> ByteString

txInfoId :: TxId crypto -> TxId

transStakeCred :: Credential keyrole crypto -> Credential

transStakeReference :: StakeReference crypto -> Maybe StakingCredential

transCred :: Credential keyrole crypto -> Credential

transAddr :: Addr crypto -> Maybe Address

transTxOutAddr :: Era era => TxOut era -> Maybe Address

slotToPOSIXTime :: HasField "_protocolVersion" (PParams era) ProtVer => PParams era -> EpochInfo (Either Text) -> SystemStart -> SlotNo -> Either Text POSIXTime

transVITime :: HasField "_protocolVersion" (PParams era) ProtVer => PParams era -> EpochInfo (Either Text) -> SystemStart -> ValidityInterval -> Either Text POSIXTimeRange

translate a validity interval to POSIX time

txInfoIn' :: TxIn c -> TxOutRef

txInfoIn :: forall era c i. (Era era, Value era ~ Value (Crypto era), HasField "datahash" (TxOut era) (StrictMaybe (SafeHash c i))) => TxIn (Crypto era) -> TxOut era -> Maybe TxInInfo

Given a TxIn, look it up in the UTxO. If it exists, translate it and return (Just translation). If does not exist in the UTxO, return Nothing.

txInfoOut :: forall era c. (Era era, Value era ~ Value (Crypto era), HasField "datahash" (TxOut era) (StrictMaybe (DataHash c))) => TxOut era -> Maybe TxOut

Given a TxOut, translate it and return (Just transalation). It is possible the address part is a Bootstrap Address, in that case return Nothing I.e. don't include Bootstrap Addresses in the answer.

transPolicyID :: PolicyID crypto -> CurrencySymbol

transAssetName :: AssetName -> TokenName

transValue :: Value c -> Value

transDCert :: DCert c -> DCert

transWdrl :: Wdrl crypto -> Map StakingCredential Integer

getWitVKeyHash :: (Crypto crypto, Typeable kr) => WitVKey kr crypto -> PubKeyHash

transDataPair :: (DataHash c, Data era) -> (DatumHash, Datum)

transExUnits :: ExUnits -> ExBudget

exBudgetToExUnits :: ExBudget -> Maybe ExUnits

transScriptPurpose :: ScriptPurpose crypto -> ScriptPurpose

data VersionedTxInfo

Constructors

TxInfoPV1 TxInfo

TxInfoPV2 TxInfo

class ExtendedUTxO era where

Where we keep functions that differ from Era to Era but which deal with the extra things in the TxOut (Scripts, DataHash, Datum, etc)

Minimal complete definition

txInfo, inputDataHashes, txscripts, getAllowedSupplimentalDataHashes, getDatum, getTxOutDatum, allSizedOuts

Methods

txInfo :: PParams era -> Language -> EpochInfo (Either Text) -> SystemStart -> UTxO era -> Tx era -> Either (TranslationError (Crypto era)) VersionedTxInfo

inputDataHashes :: Map (ScriptHash (Crypto era)) (Script era) -> ValidatedTx era -> UTxO era -> (Set (DataHash (Crypto era)), Set (TxIn (Crypto era)))

txscripts :: UTxO era -> Tx era -> Map (ScriptHash (Crypto era)) (Script era)

getAllowedSupplimentalDataHashes :: TxBody era -> UTxO era -> Set (DataHash (Crypto era))

getDatum :: Tx era -> UTxO era -> ScriptPurpose (Crypto era) -> Maybe (Data era)

getTxOutDatum :: TxOut era -> Datum era

allOuts :: TxBody era -> [TxOut era]

allSizedOuts :: TxBody era -> [Sized (TxOut era)]

alonzoTxInfo :: forall era. (Era era, Value era ~ Value (Crypto era), HasField "wits" (Tx era) (TxWitness era), HasField "datahash" (TxOut era) (StrictMaybe (SafeHash (Crypto era) EraIndependentData)), HasField "_protocolVersion" (PParams era) ProtVer, HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "reqSignerHashes" (TxBody era) (Set (KeyHash 'Witness (Crypto era))), HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era))), HasField "wdrls" (TxBody era) (Wdrl (Crypto era)), HasField "mint" (TxBody era) (Value (Crypto era)), HasField "vldt" (TxBody era) ValidityInterval) => PParams era -> Language -> EpochInfo (Either Text) -> SystemStart -> UTxO era -> Tx era -> Either (TranslationError (Crypto era)) VersionedTxInfo

valContext :: VersionedTxInfo -> ScriptPurpose (Crypto era) -> Data era

valContext pairs transaction data with a script purpose. See figure 22 of the Alonzo specification.

data ScriptFailure

Constructors

PlutusSF Text PlutusDebug

data ScriptResult

Constructors

Passes [PlutusDebug]

Fails [PlutusDebug] (NonEmpty ScriptFailure)

scriptPass :: PlutusDebug -> ScriptResult

scriptFail :: ScriptFailure -> ScriptResult

data PlutusDebug

Constructors

PlutusDebugV1 CostModel ExUnits ShortByteString [Data] ProtVer

PlutusDebugV2 CostModel ExUnits ShortByteString [Data] ProtVer

data PlutusError

Constructors

PlutusErrorV1 EvaluationError

PlutusErrorV2 EvaluationError

data PlutusDebugInfo

Constructors

DebugSuccess ExBudget

DebugCannotDecode String

DebugInfo [Text] PlutusError PlutusDebug

DebugBadHex String

debugPlutus :: String -> PlutusDebugInfo

runPLCScript :: forall era. Show (Script era) => Proxy era -> ProtVer -> Language -> CostModel -> ShortByteString -> ExUnits -> [Data] -> ScriptResult

Run a Plutus Script, given the script and the bounds on resources it is allocated.

explainPlutusFailure :: forall era. Show (Script era) => Proxy era -> ProtVer -> Language -> ShortByteString -> EvaluationError -> [Data] -> CostModel -> ExUnits -> ScriptResult

Explain why a script might fail. Scripts come in two flavors:

with 3 data arguments [data,redeemer,context]

with 2 data arguments [redeemer,context].

It pays to decode the context data into a real context because that provides way more information. But there is no guarantee the context data really can be decoded.

validPlutusdata :: Data -> Bool

validScript :: ProtVer -> Script era -> Bool

Test that every Alonzo script represents a real Script. Run deepseq to see that there are no infinite computations and that every Plutus Script unflattens into a real PV1.Script

transProtocolVersion :: ProtVer -> ProtocolVersion

languages :: forall era. (ExtendedUTxO era, Script era ~ Script era) => Tx era -> UTxO era -> Set (ScriptHash (Crypto era)) -> Set Language

Compute the Set of Languages in an era, where Alonzo.Scripts are used