Cardano.Ledger.Alonzo.TxSeq

Description

TxSeq. This is effectively the block body, which consists of a sequence of transactions with segregated witness and metadata information.

Documentation

data AlonzoTxSeq era where

Bundled Patterns

pattern AlonzoTxSeq ∷ ∀ era. (AlonzoEraTx era, SafeToHash (TxWits era)) ⇒ StrictSeq (Tx era) → AlonzoTxSeq era

hashAlonzoTxSeq ∷ ∀ era. AlonzoTxSeq era → Hash HASH EraIndependentBlockBody

Hash a given block body