Test.Cardano.Ledger.Alonzo.Arbitrary

Documentation

alwaysSucceeds ∷ Language → Natural → AlonzoScript era

alwaysFails ∷ Language → Natural → AlonzoScript era

costModelParamsCount ∷ Language → Int

newtype FlexibleCostModels

This Arbitrary instance assumes the flexible deserialization scheme of CostModels starting at version 9.

Constructors

FlexibleCostModels CostModels

Orphan instances

Arbitrary CoinPerWord

Instance details

Methods

arbitrary ∷ Gen CoinPerWord

shrink ∷ CoinPerWord → [CoinPerWord]

Arbitrary AlonzoGenesis

Instance details

Methods

arbitrary ∷ Gen AlonzoGenesis

shrink ∷ AlonzoGenesis → [AlonzoGenesis]

Arbitrary OrdExUnits

Instance details

Methods

arbitrary ∷ Gen OrdExUnits

shrink ∷ OrdExUnits → [OrdExUnits]

Arbitrary FailureDescription

Instance details

Methods

arbitrary ∷ Gen FailureDescription

shrink ∷ FailureDescription → [FailureDescription]

Arbitrary TagMismatchDescription

Instance details

Methods

arbitrary ∷ Gen TagMismatchDescription

shrink ∷ TagMismatchDescription → [TagMismatchDescription]

Arbitrary CostModel

Instance details

Methods

arbitrary ∷ Gen CostModel

shrink ∷ CostModel → [CostModel]

Arbitrary CostModels
This Arbitrary instance assumes the inflexible deserialization scheme prior to version 9.

Instance details

Methods

arbitrary ∷ Gen CostModels

shrink ∷ CostModels → [CostModels]

Arbitrary ExUnits

Instance details

Methods

arbitrary ∷ Gen ExUnits

shrink ∷ ExUnits → [ExUnits]

Arbitrary Prices

Instance details

Methods

arbitrary ∷ Gen Prices

shrink ∷ Prices → [Prices]

Arbitrary Tag

Instance details

Methods

arbitrary ∷ Gen Tag

shrink ∷ Tag → [Tag]

Arbitrary IsValid

Instance details

Methods

arbitrary ∷ Gen IsValid

shrink ∷ IsValid → [IsValid]

Arbitrary RdmrPtr

Instance details

Methods

arbitrary ∷ Gen RdmrPtr

shrink ∷ RdmrPtr → [RdmrPtr]

Arbitrary Language

Instance details

Methods

arbitrary ∷ Gen Language

shrink ∷ Language → [Language]

Arbitrary Data

Instance details

Methods

arbitrary ∷ Gen Data

shrink ∷ Data → [Data]

(EraTxOut era, Era era, Arbitrary (Value era), Arbitrary (TxOut era), Arbitrary (PredicateFailure (EraRule "UTXOS" era))) ⇒ Arbitrary (AlonzoUtxoPredFailure era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoUtxoPredFailure era)

shrink ∷ AlonzoUtxoPredFailure era → [AlonzoUtxoPredFailure era]

(Era era, Arbitrary (PPUPPredFailure era)) ⇒ Arbitrary (AlonzoUtxosPredFailure era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoUtxosPredFailure era)

shrink ∷ AlonzoUtxosPredFailure era → [AlonzoUtxosPredFailure era]

(Era era, Arbitrary (PredicateFailure (EraRule "UTXO" era)), Arbitrary (ShelleyUtxowPredFailure era)) ⇒ Arbitrary (AlonzoUtxowPredFailure era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoUtxowPredFailure era)

shrink ∷ AlonzoUtxowPredFailure era → [AlonzoUtxowPredFailure era]

Era era ⇒ Arbitrary (AlonzoScript era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoScript era)

shrink ∷ AlonzoScript era → [AlonzoScript era]

Era era ⇒ Arbitrary (BinaryData era)

Instance details

Methods

arbitrary ∷ Gen (BinaryData era)

shrink ∷ BinaryData era → [BinaryData era]

Era era ⇒ Arbitrary (Data era)

Instance details

Methods

arbitrary ∷ Gen (Data era)

shrink ∷ Data era → [Data era]

Era era ⇒ Arbitrary (Datum era)

Instance details

Methods

arbitrary ∷ Gen (Datum era)

shrink ∷ Datum era → [Datum era]

(Arbitrary (TxBody era), Arbitrary (TxWits era), Arbitrary (TxAuxData era)) ⇒ Arbitrary (AlonzoTx era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoTx era)

shrink ∷ AlonzoTx era → [AlonzoTx era]

(AlonzoEraPParams era, Arbitrary (PParams era)) ⇒ Arbitrary (ScriptIntegrity era)

Instance details

Methods

arbitrary ∷ Gen (ScriptIntegrity era)

shrink ∷ ScriptIntegrity era → [ScriptIntegrity era]

Crypto c ⇒ Arbitrary (ScriptPurpose c)

Instance details

Methods

arbitrary ∷ Gen (ScriptPurpose c)

shrink ∷ ScriptPurpose c → [ScriptPurpose c]

(Arbitrary (Script era), Era era) ⇒ Arbitrary (AlonzoTxAuxData era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoTxAuxData era)

shrink ∷ AlonzoTxAuxData era → [AlonzoTxAuxData era]

(EraTxOut era, Arbitrary (TxOut era), Arbitrary (PParamsHKD StrictMaybe era)) ⇒ Arbitrary (AlonzoTxBody era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoTxBody era)

shrink ∷ AlonzoTxBody era → [AlonzoTxBody era]

(EraTxOut era, Arbitrary (Value era)) ⇒ Arbitrary (AlonzoTxOut era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoTxOut era)

shrink ∷ AlonzoTxOut era → [AlonzoTxOut era]

(Era era, Arbitrary (Script era), AlonzoScript era ~ Script era, EraScript era) ⇒ Arbitrary (AlonzoTxWits era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoTxWits era)

shrink ∷ AlonzoTxWits era → [AlonzoTxWits era]

Era era ⇒ Arbitrary (Redeemers era)

Instance details

Methods

arbitrary ∷ Gen (Redeemers era)

shrink ∷ Redeemers era → [Redeemers era]

Arbitrary (AlonzoPParams Identity era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoPParams Identity era)

shrink ∷ AlonzoPParams Identity era → [AlonzoPParams Identity era]

Arbitrary (AlonzoPParams StrictMaybe era)

Instance details

Methods

arbitrary ∷ Gen (AlonzoPParams StrictMaybe era)

shrink ∷ AlonzoPParams StrictMaybe era → [AlonzoPParams StrictMaybe era]