Cardano.Ledger.Alonzo.Plutus.TxInfo

Documentation

data AlonzoContextError era

Constructors

TranslationLogicMissingInput !(TxIn (EraCrypto era))

TimeTranslationPastHorizon !Text

data TxOutSource c

A transaction output can be translated because it is a newly created output, or because it is the output which is connected to a transaction input being spent.

Constructors

TxOutFromInput !(TxIn c)

TxOutFromOutput !TxIx

transLookupTxOut ∷ ∀ era a. Inject (AlonzoContextError era) a ⇒ UTxO era → TxIn (EraCrypto era) → Either a (TxOut era)

transTxOut ∷ (Value era ~ MaryValue c, AlonzoEraTxOut era) ⇒ TxOut era → Maybe TxOut

Translate a TxOut. Returns Nothing if a Byron address is present in the TxOut.

transValidityInterval ∷ ∀ proxy era a. Inject (AlonzoContextError era) a ⇒ proxy era → ProtVer → EpochInfo (Either Text) → SystemStart → ValidityInterval → Either a POSIXTimeRange

Translate a validity interval to POSIX time

transPolicyID ∷ PolicyID c → CurrencySymbol

transAssetName ∷ AssetName → TokenName

transMultiAsset ∷ MultiAsset c → Value

transMintValue ∷ MultiAsset c → Value

Hysterical raisins:

Previously transaction body contained a mint field with MaryValue instead of a MultiAsset, which has changed since then to just MultiAsset (because minting ADA makes no sense). However, if we don't preserve previous translation, scripts that previously succeeded will fail.

transValue ∷ MaryValue c → Value

transWithdrawals ∷ Withdrawals c → Map StakingCredential Integer

transDataPair ∷ (DataHash c, Data era) → (DatumHash, Datum)

transTxCert ∷ (ShelleyEraTxCert era, ProtVerAtMost era 8) ⇒ TxCert era → DCert

transTxCertCommon ∷ ShelleyEraTxCert era ⇒ TxCert era → Maybe DCert

Just like transTxCert, but do not translate certificates that were deprecated in Conway

transPlutusPurpose ∷ (EraPlutusTxInfo l era, PlutusTxCert l ~ DCert) ⇒ proxy l → ProtVer → AlonzoPlutusPurpose AsItem era → Either (ContextError era) ScriptPurpose

transTxBodyId ∷ EraTxBody era ⇒ TxBody era → TxId

transTxBodyCerts ∷ (EraPlutusTxInfo l era, EraTxBody era) ⇒ proxy l → ProtVer → TxBody era → Either (ContextError era) [PlutusTxCert l]

Translate all TxCerts from within a TxBody

transTxBodyWithdrawals ∷ EraTxBody era ⇒ TxBody era → [(StakingCredential, Integer)]

Translate all Withdrawals from within a TxBody

transTxBodyReqSignerHashes ∷ AlonzoEraTxBody era ⇒ TxBody era → [PubKeyHash]

Translate all required signers produced by reqSignerHashesTxBodyLs from within a TxBody

transTxWitsDatums ∷ AlonzoEraTxWits era ⇒ TxWits era → [(DatumHash, Datum)]

Translate all TxDatss from within TxWits

LgacyPlutusArgs helpers

toPlutusV1Args ∷ EraPlutusTxInfo 'PlutusV1 era ⇒ proxy 'PlutusV1 → ProtVer → TxInfo → PlutusPurpose AsIxItem era → Maybe (Data era) → Data era → Either (ContextError era) (PlutusArgs 'PlutusV1)

toLegacyPlutusArgs ∷ EraPlutusTxInfo l era ⇒ proxy l → ProtVer → (PlutusScriptPurpose l → PlutusScriptContext l) → PlutusPurpose AsIxItem era → Maybe (Data era) → Data era → Either (ContextError era) (LegacyPlutusArgs l)

Orphan instances

Crypto c ⇒ EraPlutusTxInfo 'PlutusV1 (AlonzoEra c)

Instance details

Methods

toPlutusTxCert ∷ proxy 'PlutusV1 → ProtVer → TxCert (AlonzoEra c) → Either (ContextError (AlonzoEra c)) (PlutusTxCert 'PlutusV1)

toPlutusScriptPurpose ∷ proxy 'PlutusV1 → ProtVer → PlutusPurpose AsIxItem (AlonzoEra c) → Either (ContextError (AlonzoEra c)) (PlutusScriptPurpose 'PlutusV1)

toPlutusTxInfo ∷ proxy 'PlutusV1 → LedgerTxInfo (AlonzoEra c) → Either (ContextError (AlonzoEra c)) (PlutusTxInfo 'PlutusV1)

toPlutusArgs ∷ proxy 'PlutusV1 → ProtVer → PlutusTxInfo 'PlutusV1 → PlutusPurpose AsIxItem (AlonzoEra c) → Maybe (Data (AlonzoEra c)) → Data (AlonzoEra c) → Either (ContextError (AlonzoEra c)) (PlutusArgs 'PlutusV1)

Crypto c ⇒ EraPlutusContext (AlonzoEra c)

Instance details

Associated Types

type ContextError (AlonzoEra c) = (r ∷ Type)

Methods

mkPlutusWithContext ∷ PlutusScript (AlonzoEra c) → ScriptHash (EraCrypto (AlonzoEra c)) → PlutusPurpose AsIxItem (AlonzoEra c) → LedgerTxInfo (AlonzoEra c) → (Data (AlonzoEra c), ExUnits) → CostModel → Either (ContextError (AlonzoEra c)) (PlutusWithContext (EraCrypto (AlonzoEra c)))