Cardano.Ledger.Alonzo.PlutusScriptApi

Documentation

getSpendingTxIn ∷ ScriptPurpose c → Maybe (TxIn c)

Only the Spending ScriptPurpose contains TxIn

getDatumAlonzo ∷ (AlonzoEraTxOut era, EraTx era, TxWits era ~ AlonzoTxWits era) ⇒ Tx era → UTxO era → ScriptPurpose (EraCrypto era) → Maybe (Data era)

Get the Data associated with a ScriptPurpose. Only the Spending ScriptPurpose contains Data. The null list is returned for the other kinds.

evalScripts ∷ ∀ era. (EraTx era, Script era ~ AlonzoScript era) ⇒ ProtVer → Tx era → [(ShortByteString, Language, [Data era], ExUnits, CostModel)] → ScriptResult

evaluate a list of scripts, All scripts in the list must be True. There are two kinds of scripts, evaluate each kind using the appropriate mechanism.

scriptsNeeded ∷ (EraTx era, EraUTxO era, ScriptsNeeded era ~ [(ScriptPurpose (EraCrypto era), ScriptHash (EraCrypto era))]) ⇒ UTxO era → Tx era → [(ScriptPurpose (EraCrypto era), ScriptHash (EraCrypto era))]

Deprecated: In favor of getScritpsNeeded

scriptsNeededFromBody ∷ ∀ era. (EraUTxO era, ScriptsNeeded era ~ [(ScriptPurpose (EraCrypto era), ScriptHash (EraCrypto era))]) ⇒ UTxO era → TxBody era → [(ScriptPurpose (EraCrypto era), ScriptHash (EraCrypto era))]

Deprecated: In favor of getScritpsNeeded

language ∷ AlonzoScript era → Maybe Language

data CollectError c

When collecting inputs for twophase scripts, 3 things can go wrong.

Constructors

NoRedeemer !(ScriptPurpose c)

NoWitness !(ScriptHash c)

NoCostModel !Language

BadTranslation !(TranslationError c)

collectTwoPhaseScriptInputs ∷ ∀ era. (EraTx era, MaryEraTxBody era, AlonzoEraTxWits era, EraUTxO era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, ExtendedUTxO era, Script era ~ AlonzoScript era, AlonzoEraPParams era) ⇒ EpochInfo (Either Text) → SystemStart → PParams era → Tx era → UTxO era → Either [CollectError (EraCrypto era)] [(ShortByteString, Language, [Data era], ExUnits, CostModel)]

Collect the inputs for twophase scripts. If any script can't find ist data return a list of CollectError, if all goes well return a list of quadruples with the inputs. Previous PredicateFailure tests should ensure we find Data for every script, BUT the consequences of not finding Data means scripts can get dropped, so things might validate that shouldn't. So we double check that every Script has its Data, and if that is not the case, a PredicateFailure is raised in the Utxos rule.

knownToNotBe1Phase ∷ Map (ScriptHash (EraCrypto era)) (AlonzoScript era) → (ScriptPurpose (EraCrypto era), ScriptHash (EraCrypto era)) → Maybe (ScriptPurpose (EraCrypto era), Language, ShortByteString)