Cardano.Ledger.Alonzo.Rules.Utxos

Documentation

data UTXOS era

data UtxosPredicateFailure era

Constructors

ValidationTagMismatch IsValid TagMismatchDescription
The isValid tag on the transaction is incorrect. The tag given here is that provided on the transaction (whereas evaluation of the scripts gives the opposite.). The Text tries to explain why it failed.

CollectErrors [CollectError (Crypto era)]
We could not find all the necessary inputs for a Plutus Script. Previous PredicateFailure tests should make this impossible, but the consequences of not detecting this means scripts get dropped, so things might validate that shouldn't. So we double check in the function collectTwoPhaseScriptInputs, it should find data for every Script.

UpdateFailure (PredicateFailure (EraRule "PPUP" era))

lbl2Phase :: Label

Indicates that this check depends only upon the signal to the transition, not the state or environment.

data TagMismatchDescription

Constructors

PassedUnexpectedly

FailedUnexpectedly (NonEmpty FailureDescription)

validBegin :: String

validEnd :: String

invalidBegin :: String

invalidEnd :: String

data UtxosEvent era

Constructors

AlonzoPpupToUtxosEvent (Event (EraRule "PPUP" era))

SuccessfulPlutusScriptsEvent (NonEmpty PlutusDebug)

FailedPlutusScriptsEvent (NonEmpty PlutusDebug)

when2Phase :: Rule sts ctx () -> Rule sts ctx ()

Construct a 2-phase predicate check.

Note that 2-phase predicate checks are by definition static.

type ConcreteAlonzo era = (Script era ~ Script era, Value era ~ Value (Crypto era), TxBody era ~ TxBody era, PParams era ~ PParams era, PParamsDelta era ~ PParamsUpdate era, TxOut era ~ TxOut era, Witnesses era ~ TxWitness era, Tx era ~ ValidatedTx era)

data FailureDescription

Constructors

PlutusFailure Text ByteString

scriptFailuresToPredicateFailure :: NonEmpty ScriptFailure -> NonEmpty FailureDescription

scriptFailuresToPlutusDebug :: NonEmpty ScriptFailure -> NonEmpty PlutusDebug