Cardano.Ledger.Alonzo.Rules.Utxow

Documentation

data UtxowPredicateFail era

The Predicate failure type in the Alonzo Era. It embeds the Predicate failure type of the Shelley Era, as they share some failure modes.

Constructors

WrappedShelleyEraFailure !(UtxowPredicateFailure era)

MissingRedeemers ![(ScriptPurpose (Crypto era), ScriptHash (Crypto era))]
List of scripts for which no redeemers were supplied

MissingRequiredDatums

Fields

!(Set (DataHash (Crypto era)))

Set of missing data hashes

!(Set (DataHash (Crypto era)))

Set of received data hashes

NonOutputSupplimentaryDatums

Fields

!(Set (DataHash (Crypto era)))

Set of unallowed data hashes

!(Set (DataHash (Crypto era)))

Set of acceptable supplimental data hashes

PPViewHashesDontMatch

Fields

!(StrictMaybe (ScriptIntegrityHash (Crypto era)))

The PPHash in the TxBody

!(StrictMaybe (ScriptIntegrityHash (Crypto era)))

Computed from the current Protocol Parameters

MissingRequiredSigners (Set (KeyHash 'Witness (Crypto era)))
Set of witnesses which were needed and not supplied

UnspendableUTxONoDatumHash (Set (TxIn (Crypto era)))
Set of transaction inputs that are TwoPhase scripts, and should have a DataHash but don't

ExtraRedeemers ![RdmrPtr]
List of redeemers not needed

newtype AlonzoEvent era

Constructors

WrappedShelleyEraEvent (UtxowEvent era)

encodePredFail :: (Era era, ToCBOR (PredicateFailure (EraRule "UTXO" era)), Typeable (Script era), Typeable (AuxiliaryData era)) => UtxowPredicateFail era -> Encode 'Open (UtxowPredicateFail era)

decodePredFail :: (Era era, FromCBOR (PredicateFailure (EraRule "UTXO" era)), Typeable (Script era), Typeable (AuxiliaryData era)) => Word -> Decode 'Open (UtxowPredicateFail era)

langsUsed :: forall era. (Script era ~ Script era, ValidateScript era) => Map (ScriptHash (Crypto era)) (Script era) -> Set Language

given the "txscripts" field of the Witnesses, compute the set of languages used in a transaction

missingRequiredDatums :: forall era. (ValidateScript era, Script era ~ Script era, ExtendedUTxO era) => Map (ScriptHash (Crypto era)) (Script era) -> UTxO era -> ValidatedTx era -> TxBody era -> Test (UtxowPredicateFail era)

hasExactSetOfRedeemers :: forall era. (Era era, ValidateScript era, ExtendedUTxO era, Script era ~ Script era, Tx era ~ ValidatedTx era, HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era))), HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "wdrls" (TxBody era) (Wdrl (Crypto era))) => UTxO era -> Tx era -> TxBody era -> Test (UtxowPredicateFail era)

requiredSignersAreWitnessed :: forall era. HasField "reqSignerHashes" (TxBody era) (Set (KeyHash 'Witness (Crypto era))) => TxBody era -> WitHashes (Crypto era) -> Test (UtxowPredicateFail era)

ppViewHashesMatch :: forall era. (ValidateScript era, ExtendedUTxO era, Script era ~ Script era, Tx era ~ ValidatedTx era, HasField "scriptIntegrityHash" (TxBody era) (StrictMaybe (ScriptIntegrityHash (Crypto era))), HasField "_costmdls" (PParams era) CostModels) => Tx era -> TxBody era -> PParams era -> UTxO era -> Set (ScriptHash (Crypto era)) -> Test (UtxowPredicateFail era)

alonzoStyleWitness :: forall era. (ValidateScript era, ValidateAuxiliaryData era (Crypto era), ExtendedUTxO era, ConcreteAlonzo era, Tx era ~ ValidatedTx era, Witnesses era ~ TxWitness era, Signable (DSIGN (Crypto era)) (Hash (HASH (Crypto era)) EraIndependentTxBody), Embed (EraRule "UTXO" era) (AlonzoUTXOW era), Environment (EraRule "UTXO" era) ~ UtxoEnv era, State (EraRule "UTXO" era) ~ UTxOState era, Signal (EraRule "UTXO" era) ~ ValidatedTx era) => TransitionRule (AlonzoUTXOW era)

A very specialized transitionRule function for the Alonzo Era.

witsVKeyNeeded :: forall era tx. (Era era, HasField "body" tx (TxBody era), HasField "wdrls" (TxBody era) (Wdrl (Crypto era)), HasField "certs" (TxBody era) (StrictSeq (DCert (Crypto era))), HasField "inputs" (TxBody era) (Set (TxIn (Crypto era))), HasField "collateral" (TxBody era) (Set (TxIn (Crypto era))), HasField "update" (TxBody era) (StrictMaybe (Update era))) => UTxO era -> tx -> GenDelegs (Crypto era) -> WitHashes (Crypto era)

Collect the set of hashes of keys that needs to sign a given transaction. This set consists of the txin owners, certificate authors, and withdrawal reward accounts.

Compared to pre-Alonzo eras, we additionally gather the certificates required to authenticate collateral witnesses.

extSymmetricDifference :: Ord k => [a] -> (a -> k) -> [b] -> (b -> k) -> ([a], [b])

data AlonzoUTXOW era