Cardano.Ledger.Alonzo.Rules.Utxow

Documentation

data AlonzoUTXOW era

newtype AlonzoUtxowEvent era

Constructors

WrappedShelleyEraEvent (ShelleyUtxowEvent era)

data AlonzoUtxowPredFailure era

The Predicate failure type in the Alonzo Era. It embeds the Predicate failure type of the Shelley Era, as they share some failure modes.

Constructors

ShelleyInAlonzoUtxowPredFailure !(ShelleyUtxowPredFailure era)

MissingRedeemers ![(ScriptPurpose (EraCrypto era), ScriptHash (EraCrypto era))]
List of scripts for which no redeemers were supplied

MissingRequiredDatums

Fields

!(Set (DataHash (EraCrypto era)))

Set of missing data hashes

!(Set (DataHash (EraCrypto era)))

Set of received data hashes

NonOutputSupplimentaryDatums

Fields

!(Set (DataHash (EraCrypto era)))

Set of unallowed data hashes

!(Set (DataHash (EraCrypto era)))

Set of acceptable supplimental data hashes

PPViewHashesDontMatch

Fields

!(StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

The PPHash in the TxBody

!(StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

Computed from the current Protocol Parameters

MissingRequiredSigners (Set (KeyHash 'Witness (EraCrypto era)))
Set of witnesses which were needed and not supplied

UnspendableUTxONoDatumHash (Set (TxIn (EraCrypto era)))
Set of transaction inputs that are TwoPhase scripts, and should have a DataHash but don't

ExtraRedeemers ![RdmrPtr]
List of redeemers not needed

hasExactSetOfRedeemers ∷ ∀ era. (AlonzoEraTx era, ExtendedUTxO era, Script era ~ AlonzoScript era) ⇒ UTxO era → Tx era → AlonzoScriptsNeeded era → Test (AlonzoUtxowPredFailure era)

missingRequiredDatums ∷ ∀ era. (AlonzoEraTx era, ExtendedUTxO era) ⇒ Map (ScriptHash (EraCrypto era)) (Script era) → UTxO era → Tx era → Test (AlonzoUtxowPredFailure era)

ppViewHashesMatch ∷ ∀ era. (AlonzoEraTx era, ExtendedUTxO era, Script era ~ AlonzoScript era) ⇒ Tx era → PParams era → UTxO era → Set (ScriptHash (EraCrypto era)) → Test (AlonzoUtxowPredFailure era)

requiredSignersAreWitnessed ∷ AlonzoEraTxBody era ⇒ TxBody era → Set (KeyHash 'Witness (EraCrypto era)) → Test (AlonzoUtxowPredFailure era)

witsVKeyNeeded ∷ ∀ era. (EraTx era, AlonzoEraTxBody era) ⇒ UTxO era → Tx era → GenDelegs (EraCrypto era) → Set (KeyHash 'Witness (EraCrypto era))

Collect the set of hashes of keys that needs to sign a given transaction. This set consists of the txin owners, certificate authors, and withdrawal reward accounts.

Compared to pre-Alonzo eras, we additionally gather the certificates required to authenticate collateral witnesses.

Orphan instances

(AlonzoEraTx era, EraTxAuxData era, ExtendedUTxO era, EraUTxO era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, Signable (DSIGN (EraCrypto era)) (Hash (HASH (EraCrypto era)) EraIndependentTxBody), Script era ~ AlonzoScript era, Embed (EraRule "UTXO" era) (AlonzoUTXOW era), Environment (EraRule "UTXO" era) ~ UtxoEnv era, State (EraRule "UTXO" era) ~ UTxOState era, Signal (EraRule "UTXO" era) ~ Tx era, ProtVerAtMost era 8) ⇒ STS (AlonzoUTXOW era)

Instance details

Associated Types

type State (AlonzoUTXOW era)

type Signal (AlonzoUTXOW era)

type Environment (AlonzoUTXOW era)

type BaseM (AlonzoUTXOW era) ∷ Type → Type

type Event (AlonzoUTXOW era)

type PredicateFailure (AlonzoUTXOW era)

Methods

initialRules ∷ [InitialRule (AlonzoUTXOW era)]

transitionRules ∷ [TransitionRule (AlonzoUTXOW era)]

assertions ∷ [Assertion (AlonzoUTXOW era)]

renderAssertionViolation ∷ AssertionViolation (AlonzoUTXOW era) → String

(Era era, STS (AlonzoUTXO era), PredicateFailure (EraRule "UTXO" era) ~ AlonzoUtxoPredFailure era, Event (EraRule "UTXO" era) ~ AlonzoUtxoEvent era, BaseM (AlonzoUTXOW era) ~ ShelleyBase, PredicateFailure (AlonzoUTXOW era) ~ AlonzoUtxowPredFailure era, Event (AlonzoUTXOW era) ~ AlonzoUtxowEvent era) ⇒ Embed (AlonzoUTXO era) (AlonzoUTXOW era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (AlonzoUTXO era) → PredicateFailure (AlonzoUTXOW era)

wrapEvent ∷ Event (AlonzoUTXO era) → Event (AlonzoUTXOW era)