Cardano.Ledger.Alonzo.Scripts

Documentation

data Tag

Marker indicating the part of a transaction for which this script is acting as a validator.

Constructors

Spend
Validates spending a script-locked UTxO

Mint
Validates minting new tokens

Cert
Validates certificate transactions

Rewrd
Validates withdrawl from a reward account

data Script era

Scripts in the Alonzo Era, Either a Timelock script or a Plutus script.

Constructors

TimelockScript (Timelock (Crypto era))

PlutusScript Language ShortByteString

txscriptfee :: Prices -> ExUnits -> Coin

Compute the cost of a script based upon prices and the number of execution units.

isPlutusScript :: Script era -> Bool

pointWiseExUnits :: (Natural -> Natural -> Bool) -> ExUnits -> ExUnits -> Bool

It is deliberate that there is no Ord instance for ExUnits. Use this function to compare if one ExUnit is pointwise compareable to another.

Cost Model

data CostModel

A language dependent cost model for the Plutus evaluator. Note that the EvaluationContext is entirely dependent on the cost model parameters (ie the Map Text Integer) and that this type uses the smart constructor mkCostModel to hide the evaluation context.

mkCostModel :: Language -> Map Text Integer -> Either CostModelApplyError CostModel

Convert cost model parameters to a cost model, making use of the conversion function mkEvaluationContext from the Plutus API.

getCostModelLanguage :: CostModel -> Language

getCostModelParams :: CostModel -> Map Text Integer

getEvaluationContext :: CostModel -> EvaluationContext

newtype ExUnits

This newtype wrapper of ExUnits' is used to hide an implementation detail inside the ExUnits pattern.

Constructors

WrapExUnits

Fields

unWrapExUnits :: ExUnits' Natural

Bundled Patterns

pattern ExUnits :: Natural -> Natural -> ExUnits
Arbitrary execution unit in which we measure the cost of scripts in terms of space in memory and execution time.

This pattern hides the fact that ExUnits' is parametric in the underlying type. The ledger itself uses ExUnits Natural' exclusively.

We would have preferred to use a type alias for ExUnits Natural', but this is not possible: https://gitlab.haskell.org/ghc/ghc/-/issues/19507.

data ExUnits' a

Arbitrary execution unit in which we measure the cost of scripts in terms of space in memory and execution time.

The ledger itself uses ExUnits Natural' exclusively, but the flexibility here alows the consensus layer to translate the execution units into something equivalent to 'ExUnits (Inf Natural)'. This is needed in order to provide a BoundedMeasure instance, which itself is needed for the alonzo instance of TxLimits (in consensus).

data Prices

Prices per execution unit

Constructors

Prices

Fields

prMem :: !NonNegativeInterval

prSteps :: !NonNegativeInterval

hashCostModel :: forall e. Era e => Proxy e -> CostModel -> SafeHash (Crypto e) CostModel

assertWellFormedCostModelParams :: MonadError CostModelApplyError m => CostModelParams -> m ()

Comparably expensive to mkEvaluationContext, so it should only be used sparingly.

decodeCostModelMap :: Decoder s (Map Language CostModel)

decodeCostModel :: Language -> Decoder s CostModel

newtype CostModels

Constructors

CostModels

Fields

unCostModels :: Map Language CostModel

data CostModelApplyError

The type of errors that applyParams can throw.

Constructors

CMUnknownParamError Text
a costmodel parameter with the give name does not exist in the costmodel to be applied upon

CMInternalReadError
internal error when we are transforming the applyParams' input to json (should not happen)

CMInternalWriteError String
internal error when we are transforming the applied params from json with given jsonstring error (should not happen)