Cardano.Ledger.Alonzo.Scripts

Documentation

data Tag

Marker indicating the part of a transaction for which this script is acting as a validator.

Constructors

Spend
Validates spending a script-locked UTxO

Mint
Validates minting new tokens

Cert
Validates certificate transactions

Rewrd
Validates withdrawal from a reward account

newtype BinaryPlutus

Binary representation of a Plutus script.

Constructors

BinaryPlutus

Fields

unBinaryPlutus ∷ ShortByteString

data AlonzoScript era

Scripts in the Alonzo Era, Either a Timelock script or a Plutus script.

Constructors

TimelockScript !(Timelock era)

PlutusScript !Language !ShortByteString

type family Script era = (r ∷ Type) | r → era

Scripts which may lock transaction outputs in this era

txscriptfee ∷ Prices → ExUnits → Coin

Compute the cost of a script based upon prices and the number of execution units.

isPlutusScript ∷ AlonzoScript era → Bool

pointWiseExUnits ∷ (Natural → Natural → Bool) → ExUnits → ExUnits → Bool

It is deliberate that there is no Ord instance for ExUnits. Use this function to compare if one ExUnit is pointwise compareable to another. In case when Ord instance like comparison is necessary you can use zipSemiExUnits compare

zipSemiExUnits ∷ Semigroup a ⇒ (Natural → Natural → a) → ExUnits → ExUnits → a

Pointwise combine units into a semigroup and mappened the results.

validScript ∷ ProtVer → AlonzoScript era → Bool

Test that every Alonzo script represents a real Script. Run deepseq to see that there are no infinite computations and that every Plutus Script unflattens into a real PV1.Script

transProtocolVersion ∷ ProtVer → ProtocolVersion

eqAlonzoScriptRaw ∷ AlonzoScript era → AlonzoScript era → Bool

Check the equality of two underlying types, while ignoring their binary representation, which Eq instance normally does. This is used for testing.

Cost Model

data CostModel

A language dependent cost model for the Plutus evaluator. Note that the EvaluationContext is entirely dependent on the cost model parameters (ie the Map Text Integer) and that this type uses the smart constructor mkCostModel to hide the evaluation context.

newtype CostModelError

See CostModels for an explanation of how CostModelError is used.

Constructors

CostModelError CostModelApplyError

emptyCostModels ∷ CostModels

mkCostModel ∷ Language → [Integer] → Either CostModelApplyError CostModel

Convert cost model parameters to a cost model, making use of the conversion function mkEvaluationContext from the Plutus API.

mkCostModelsLenient ∷ Map Word8 [Integer] → CostModels

encodeCostModel ∷ CostModel → Encoding

Encoding for the CostModel. Important to note that it differs from Encoding used by getLanguageView

getCostModelLanguage ∷ CostModel → Language

getCostModelParams ∷ CostModel → [Integer]

getEvaluationContext ∷ CostModel → EvaluationContext

costModelParamNames ∷ Language → [Text]

costModelToMap ∷ CostModel → Map Text Integer

costModelFromMap ∷ MonadFail m ⇒ Language → Map Text Integer → m CostModel

newtype ExUnits

This newtype wrapper of ExUnits' is used to hide an implementation detail inside the ExUnits pattern.

Constructors

WrapExUnits

Fields

unWrapExUnits ∷ ExUnits' Natural

Bundled Patterns

pattern ExUnits ∷ Natural → Natural → ExUnits
Arbitrary execution unit in which we measure the cost of scripts in terms of space in memory and execution time.

This pattern hides the fact that ExUnits' is parametric in the underlying type. The ledger itself uses ExUnits Natural' exclusively.

We would have preferred to use a type alias for ExUnits Natural', but this is not possible: https://gitlab.haskell.org/ghc/ghc/-/issues/19507.

data ExUnits' a

Arbitrary execution unit in which we measure the cost of scripts in terms of space in memory and execution time.

The ledger itself uses ExUnits Natural' exclusively, but the flexibility here allows the consensus layer to translate the execution units into something equivalent to 'ExUnits (Inf Natural)'. This is needed in order to provide a BoundedMeasure instance, which itself is needed for the alonzo instance of TxLimits (in consensus).

data Prices

Prices per execution unit

Constructors

Prices

Fields

prMem ∷ !NonNegativeInterval

prSteps ∷ !NonNegativeInterval

decodeCostModelFailHard ∷ Language → Decoder s CostModel

data CostModels

For a known version of Plutus, attempting to construct a cost model with too few parameters (depending on the version) will result in an error. CostModelApplyError exists to collect these errors in the CostModels type. The CostModels type itself needs to be flexible enough to accept any map of Word8 to '[Integer]', so that cost models can be placed in the protocol parameters ahead of changes to the Plutus evaluation context. In this way, serializing a cost model, updating software, and deserializing can result in errors going away.

Additionally, CostModels needs to be able to store cost models for future version of Plutus, which we cannot yet even validate. These are stored in invalidCostModels.

Constructors

CostModels

Fields

costModelsValid ∷ !(Map Language CostModel)

costModelsErrors ∷ !(Map Language CostModelError)

costModelsUnknown ∷ !(Map Word8 [Integer])

data CostModelApplyError

A fatal error when trying to create a cost given some plain costmodel parameters.

Constructors

CMUnknownParamError !Text
a costmodel parameter with the give name does not exist in the costmodel to be applied upon

CMInternalReadError
internal error when we are transforming the applyParams' input to json (should not happen)

CMInternalWriteError !String
internal error when we are transforming the applied params from json with given jsonstring error (should not happen)

CMTooFewParamsError
See Note [Cost model parameters from the ledger's point of view]

Fields

cmTooFewExpected ∷ !Int

cmTooFewActual ∷ !Int

Orphan instances

NFData CostModelApplyError

Instance details

Methods

rnf ∷ CostModelApplyError → ()

NoThunks CostModelApplyError

Instance details

Methods

noThunks ∷ Context → CostModelApplyError → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → CostModelApplyError → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy CostModelApplyError → String

Crypto c ⇒ EraScript (AlonzoEra c)

Instance details

Associated Types

type Script (AlonzoEra c) = (r ∷ Type)

Methods

scriptPrefixTag ∷ Script (AlonzoEra c) → ByteString

hashScript ∷ Script (AlonzoEra c) → ScriptHash (EraCrypto (AlonzoEra c))

phaseScript ∷ ∀ (phase ∷ Phase). PhaseRep phase → Script (AlonzoEra c) → Maybe (PhasedScript phase (AlonzoEra c))

isNativeScript ∷ Script (AlonzoEra c) → Bool