Cardano.Ledger.Alonzo.Tx

Description

This module exports implementations of many of the functions outlined in the Alonzo specification. The link to source of the specification https://github.com/intersectmbo/cardano-ledger/tree/master/eras/alonzo/formal-spec The most recent version of the document can be found here: https://github.com/intersectmbo/cardano-ledger/releases/latest/download/alonzo-ledger.pdf The functions can be found in Figures in that document, and sections of this code refer to those figures.

Documentation

data CostModel

A language dependent cost model for the Plutus evaluator. Note that the EvaluationContext is entirely dependent on the cost model parameters (ie the Map Text Integer) and that this type uses the smart constructor mkCostModel to hide the evaluation context.

getLanguageView ∷ AlonzoEraPParams era ⇒ PParams era → Language → LangDepView

data Data era

type DataHash = SafeHash EraIndependentData

newtype IsValid

Tag indicating whether non-native scripts in this transaction are expected to validate. This is added by the block creator when constructing the block.

Constructors

IsValid Bool

hashData ∷ Data era → DataHash

nonNativeLanguages ∷ [Language]

hashScriptIntegrity ∷ Era era ⇒ ScriptIntegrity era → ScriptIntegrityHash

data EraIndependentScriptIntegrity

data ScriptIntegrity era

A ScriptIntegrityHash is the hash of three things. The first two come from the witnesses and the last comes from the Protocol Parameters.

Constructors

ScriptIntegrity !(Redeemers era) !(TxDats era) !(Set LangDepView)

type ScriptIntegrityHash = SafeHash EraIndependentScriptIntegrity

data AlonzoTx era

Constructors

AlonzoTx

Fields

atBody ∷ !(TxBody era)

atWits ∷ !(TxWits era)

atIsValid ∷ !IsValid

atAuxData ∷ !(StrictMaybe (TxAuxData era))

data family Tx era

class (EraTx era, AlonzoEraTxBody era, AlonzoEraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTx era where

Methods

isValidTxL ∷ Lens' (Tx era) IsValid

mkBasicAlonzoTx ∷ Monoid (TxWits era) ⇒ TxBody era → AlonzoTx era

bodyAlonzoTxL ∷ Lens' (AlonzoTx era) (TxBody era)

TxBody setter and getter for AlonzoTx.

witsAlonzoTxL ∷ Lens' (AlonzoTx era) (TxWits era)

TxWits setter and getter for AlonzoTx.

auxDataAlonzoTxL ∷ Lens' (AlonzoTx era) (StrictMaybe (TxAuxData era))

TxAuxData setter and getter for AlonzoTx.

sizeAlonzoTxF ∷ ∀ era. (HasCallStack, EraTx era) ⇒ SimpleGetter (AlonzoTx era) Word32

txsize computes the length of the serialised bytes (for estimations)

isValidAlonzoTxL ∷ Lens' (AlonzoTx era) IsValid

txdats' ∷ Era era ⇒ AlonzoTxWits era → TxDats era

Deprecated: In favor of txdats

txscripts' ∷ Era era ⇒ AlonzoTxWits era → Map ScriptHash (Script era)

Deprecated: In favor of txscripts

txrdmrs ∷ AlonzoEraScript era ⇒ AlonzoTxWits era → Redeemers era

data family TxBody era

The body of a transaction.

totExUnits ∷ (EraTx era, AlonzoEraTxWits era) ⇒ Tx era → ExUnits

alonzoMinFeeTx ∷ (EraTx era, AlonzoEraTxWits era, AlonzoEraPParams era) ⇒ PParams era → Tx era → Coin

txouts ∷ EraTxBody era ⇒ TxBody era → UTxO era

Compute the transaction outputs of a transaction.

toCBORForSizeComputation ∷ (EncCBOR (TxBody era), EncCBOR (TxWits era), EncCBOR (TxAuxData era)) ⇒ AlonzoTx era → Encoding

This ensures that the size of transactions from Mary is unchanged. The individual components all store their bytes; the only work we do in this function is concatenating

toCBORForMempoolSubmission ∷ (EncCBOR (TxBody era), EncCBOR (TxWits era), EncCBOR (TxAuxData era)) ⇒ AlonzoTx era → Encoding

Encode to CBOR for the purposes of transmission from node to node, or from wallet to node.

Note that this serialisation is neither the serialisation used on-chain (where Txs are deconstructed using segwit), nor the serialisation used for computing the transaction size (which omits the IsValid field for compatibility with Mary - see toCBORForSizeComputation).

alonzoTxEqRaw ∷ AlonzoEraTx era ⇒ Tx era → Tx era → Bool

mkScriptIntegrity ∷ (AlonzoEraPParams era, AlonzoEraTxWits era, EraUTxO era) ⇒ PParams era → Tx era → ScriptsProvided era → Set ScriptHash → StrictMaybe (ScriptIntegrity era)

Orphan instances

EraTx AlonzoEra

Instance details

Associated Types

data Tx AlonzoEra

Methods

mkBasicTx ∷ TxBody AlonzoEra → Tx AlonzoEra

bodyTxL ∷ Lens' (Tx AlonzoEra) (TxBody AlonzoEra)

witsTxL ∷ Lens' (Tx AlonzoEra) (TxWits AlonzoEra)

auxDataTxL ∷ Lens' (Tx AlonzoEra) (StrictMaybe (TxAuxData AlonzoEra))

sizeTxF ∷ SimpleGetter (Tx AlonzoEra) Word32

sizeTxForFeeCalculation ∷ Tx AlonzoEra → Word32

validateNativeScript ∷ Tx AlonzoEra → NativeScript AlonzoEra → Bool

getMinFeeTx ∷ PParams AlonzoEra → Tx AlonzoEra → Int → Coin

Generic (Tx AlonzoEra)

Instance details

Associated Types

type Rep (Tx AlonzoEra) ∷ Type → Type

Methods

from ∷ Tx AlonzoEra → Rep (Tx AlonzoEra) x

to ∷ Rep (Tx AlonzoEra) x → Tx AlonzoEra

Show (Tx AlonzoEra)

Instance details

Methods

showsPrec ∷ Int → Tx AlonzoEra → ShowS

show ∷ Tx AlonzoEra → String

showList ∷ [Tx AlonzoEra] → ShowS

ToCBOR (Tx AlonzoEra)

Instance details

Methods

toCBOR ∷ Tx AlonzoEra → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy (Tx AlonzoEra) → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [Tx AlonzoEra] → Size

DecCBOR (Annotator (Tx AlonzoEra))

Instance details

Methods

decCBOR ∷ Decoder s (Annotator (Tx AlonzoEra))

dropCBOR ∷ Proxy (Annotator (Tx AlonzoEra)) → Decoder s ()

label ∷ Proxy (Annotator (Tx AlonzoEra)) → Text

EncCBOR (Tx AlonzoEra)

Instance details

Methods

encCBOR ∷ Tx AlonzoEra → Encoding

encodedSizeExpr ∷ (∀ t. EncCBOR t ⇒ Proxy t → Size) → Proxy (Tx AlonzoEra) → Size

encodedListSizeExpr ∷ (∀ t. EncCBOR t ⇒ Proxy t → Size) → Proxy [Tx AlonzoEra] → Size

EqRaw (Tx AlonzoEra)

Instance details

Methods

eqRaw ∷ Tx AlonzoEra → Tx AlonzoEra → Bool

NFData (Tx AlonzoEra)

Instance details

Methods

rnf ∷ Tx AlonzoEra → ()

Eq (Tx AlonzoEra)

Instance details

Methods

(==) ∷ Tx AlonzoEra → Tx AlonzoEra → Bool

(/=) ∷ Tx AlonzoEra → Tx AlonzoEra → Bool

NoThunks (Tx AlonzoEra)

Instance details

Methods

noThunks ∷ Context → Tx AlonzoEra → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Tx AlonzoEra → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Tx AlonzoEra) → String