Cardano.Ledger.Alonzo.TxAuxData

AlonzoTxAuxData

data AlonzoTxAuxData era where

Bundled Patterns

pattern AlonzoTxAuxData ∷ ∀ era. (HasCallStack, AlonzoEraScript era) ⇒ Map Word64 Metadatum → StrictSeq (Timelock era) → Map Language (NonEmpty PlutusBinary) → AlonzoTxAuxData era
Construct auxiliary data. Make sure not to supply plutus script versions that are not supported in this era, because it will result in a runtime exception. Use mkAlonzoTxAuxData instead if you need runtime safety guarantees.

pattern AlonzoTxAuxData' ∷ ∀ era. Map Word64 Metadatum → StrictSeq (Timelock era) → Map Language (NonEmpty PlutusBinary) → AlonzoTxAuxData era

class AllegraEraTxAuxData era ⇒ AlonzoEraTxAuxData era where

Methods

plutusScriptsTxAuxDataL ∷ Lens' (TxAuxData era) (Map Language (NonEmpty PlutusBinary))

data AlonzoTxAuxDataRaw era

mkAlonzoTxAuxData ∷ ∀ f era. (Foldable f, AlonzoEraScript era) ⇒ Map Word64 Metadatum → f (AlonzoScript era) → AlonzoTxAuxData era

Helper function that will construct Auxiliary data from Metadatum map and a list of scripts.

Note that the relative order of same type scripts will be preserved.

newtype AuxiliaryDataHash c

Constructors

AuxiliaryDataHash

Fields

unsafeAuxiliaryDataHash ∷ SafeHash c EraIndependentTxAuxData

hashAlonzoTxAuxData ∷ (HashAlgorithm (HASH c), HashAnnotated x EraIndependentTxAuxData c) ⇒ x → AuxiliaryDataHash c

validateAlonzoTxAuxData ∷ (AlonzoEraScript era, Script era ~ AlonzoScript era) ⇒ ProtVer → AuxiliaryData era → Bool

getAlonzoTxAuxDataScripts ∷ ∀ era. AlonzoEraScript era ⇒ AlonzoTxAuxData era → StrictSeq (AlonzoScript era)

translateAlonzoTxAuxData ∷ (AlonzoEraScript era1, AlonzoEraScript era2, EraCrypto era1 ~ EraCrypto era2) ⇒ AlonzoTxAuxData era1 → AlonzoTxAuxData era2

metadataAlonzoTxAuxDataL ∷ Era era ⇒ Lens' (AlonzoTxAuxData era) (Map Word64 Metadatum)

timelockScriptsAlonzoTxAuxDataL ∷ Era era ⇒ Lens' (AlonzoTxAuxData era) (StrictSeq (Timelock era))

plutusScriptsAllegraTxAuxDataL ∷ Era era ⇒ Lens' (AlonzoTxAuxData era) (Map Language (NonEmpty PlutusBinary))

Deprecated

type AuxiliaryData era = AlonzoTxAuxData era

Deprecated: Use AlonzoTxAuxData instead

Orphan instances

Crypto c ⇒ AllegraEraTxAuxData (AlonzoEra c)

Instance details

Methods

timelockScriptsTxAuxDataL ∷ Lens' (TxAuxData (AlonzoEra c)) (StrictSeq (Timelock (AlonzoEra c)))

Crypto c ⇒ EraTxAuxData (AlonzoEra c)

Instance details

Associated Types

type TxAuxData (AlonzoEra c) = (r ∷ Type)

Methods

mkBasicTxAuxData ∷ TxAuxData (AlonzoEra c)

metadataTxAuxDataL ∷ Lens' (TxAuxData (AlonzoEra c)) (Map Word64 Metadatum)

upgradeTxAuxData ∷ TxAuxData (PreviousEra (AlonzoEra c)) → TxAuxData (AlonzoEra c)

hashTxAuxData ∷ TxAuxData (AlonzoEra c) → AuxiliaryDataHash (EraCrypto (AlonzoEra c))

validateTxAuxData ∷ ProtVer → TxAuxData (AlonzoEra c) → Bool