Cardano.Ledger.Alonzo.TxAuxData

AlonzoTxAuxData

data AlonzoTxAuxData era where

Bundled Patterns

pattern AlonzoTxAuxData ∷ Era era ⇒ Map Word64 Metadatum → StrictSeq (Timelock era) → Map Language (NonEmpty BinaryPlutus) → AuxiliaryData era

mkAlonzoTxAuxData ∷ ∀ f era. (Foldable f, Era era) ⇒ Map Word64 Metadatum → f (AlonzoScript era) → AlonzoTxAuxData era

Helper function that will construct Auxiliary data from Metadatum map and a list of scripts.

Note that the relative order of same type scripts will be preserved.

newtype AuxiliaryDataHash c

Constructors

AuxiliaryDataHash

Fields

unsafeAuxiliaryDataHash ∷ SafeHash c EraIndependentTxAuxData

hashAlonzoTxAuxData ∷ (HashAlgorithm (HASH c), HashAnnotated x EraIndependentTxAuxData c) ⇒ x → AuxiliaryDataHash c

validateAlonzoTxAuxData ∷ Era era ⇒ ProtVer → AuxiliaryData era → Bool

getAlonzoTxAuxDataScripts ∷ Era era ⇒ AlonzoTxAuxData era → StrictSeq (AlonzoScript era)

Deprecated

type AuxiliaryData era = AlonzoTxAuxData era

Deprecated: Use AlonzoTxAuxData instead

Orphan instances

Crypto c ⇒ EraTxAuxData (AlonzoEra c)

Instance details

Associated Types

type TxAuxData (AlonzoEra c) = (r ∷ Type)

Methods

hashTxAuxData ∷ TxAuxData (AlonzoEra c) → AuxiliaryDataHash (EraCrypto (AlonzoEra c))

validateTxAuxData ∷ ProtVer → TxAuxData (AlonzoEra c) → Bool