Cardano.Ledger.Alonzo.TxBody

Documentation

data AlonzoTxOut era

Constructors

TxOutCompact' !CompactAddr !(CompactForm (Value era))

TxOutCompactDH' !CompactAddr !(CompactForm (Value era)) !DataHash

TxOut_AddrHash28_AdaOnly !(Credential 'Staking) !Addr28Extra !(CompactForm Coin)

TxOut_AddrHash28_AdaOnly_DataHash32 !(Credential 'Staking) !Addr28Extra !(CompactForm Coin) !DataHash32

Bundled Patterns

pattern TxOutCompact ∷ (Era era, Val (Value era), HasCallStack) ⇒ CompactAddr → CompactForm (Value era) → AlonzoTxOut era

pattern AlonzoTxOut ∷ ∀ era. (Era era, Val (Value era), HasCallStack) ⇒ Addr → Value era → StrictMaybe DataHash → AlonzoTxOut era

pattern TxOutCompactDH ∷ (Era era, Val (Value era), HasCallStack) ⇒ CompactAddr → CompactForm (Value era) → DataHash → AlonzoTxOut era

class (AlonzoEraPParams era, EraTxOut era) ⇒ AlonzoEraTxOut era where

Methods

dataHashTxOutL ∷ Lens' (TxOut era) (StrictMaybe DataHash)

datumTxOutF ∷ SimpleGetter (TxOut era) (Datum era)

data Addr28Extra

data DataHash32

data family TxBody era

The body of a transaction.

data AlonzoTxBodyRaw

Constructors

AlonzoTxBodyRaw

Fields

atbrInputs ∷ !(Set TxIn)

atbrCollateral ∷ !(Set TxIn)

atbrOutputs ∷ !(StrictSeq (TxOut AlonzoEra))

atbrCerts ∷ !(StrictSeq (TxCert AlonzoEra))

atbrWithdrawals ∷ !Withdrawals

atbrTxFee ∷ !Coin

atbrValidityInterval ∷ !ValidityInterval

atbrUpdate ∷ !(StrictMaybe (Update AlonzoEra))

atbrReqSignerHashes ∷ Set (KeyHash 'Witness)

atbrMint ∷ !MultiAsset

atbrScriptIntegrityHash ∷ !(StrictMaybe ScriptIntegrityHash)

atbrAuxDataHash ∷ !(StrictMaybe TxAuxDataHash)

atbrTxNetworkId ∷ !(StrictMaybe Network)

class (MaryEraTxBody era, AlonzoEraTxOut era) ⇒ AlonzoEraTxBody era where

Minimal complete definition

collateralInputsTxBodyL, reqSignerHashesTxBodyL, scriptIntegrityHashTxBodyL, networkIdTxBodyL, redeemerPointer, redeemerPointerInverse

Methods

collateralInputsTxBodyL ∷ Lens' (TxBody era) (Set TxIn)

reqSignerHashesTxBodyL ∷ AtMostEra "Conway" era ⇒ Lens' (TxBody era) (Set (KeyHash 'Witness))

reqSignerHashesTxBodyG ∷ SimpleGetter (TxBody era) (Set (KeyHash Witness))

default reqSignerHashesTxBodyG ∷ AtMostEra "Conway" era ⇒ SimpleGetter (TxBody era) (Set (KeyHash Witness))

scriptIntegrityHashTxBodyL ∷ Lens' (TxBody era) (StrictMaybe ScriptIntegrityHash)

networkIdTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Network)

redeemerPointer ∷ TxBody era → PlutusPurpose AsItem era → StrictMaybe (PlutusPurpose AsIx era)

This function is called rdptr in the spec. Given a TxBody and a plutus purpose with an item, we should be able to find the plutus purpose as in index

redeemerPointerInverse ∷ TxBody era → PlutusPurpose AsIx era → StrictMaybe (PlutusPurpose AsIxItem era)

This is an inverse of redeemerPointer. Given purpose as an index return it as an item.

class (ShelleyEraTxCert era, EraTxBody era, AtMostEra "Babbage" era) ⇒ ShelleyEraTxBody era where

Methods

ttlTxBodyL ∷ Lens' (TxBody era) SlotNo

updateTxBodyL ∷ Lens' (TxBody era) (StrictMaybe (Update era))

class EraTxBody era ⇒ AllegraEraTxBody era where

Methods

vldtTxBodyL ∷ Lens' (TxBody era) ValidityInterval

class AllegraEraTxBody era ⇒ MaryEraTxBody era where

Minimal complete definition

mintTxBodyL, mintedTxBodyF

Methods

mintTxBodyL ∷ Lens' (TxBody era) MultiAsset

mintedTxBodyF ∷ SimpleGetter (TxBody era) (Set PolicyID)

mintValueTxBodyF ∷ SimpleGetter (TxBody era) (Value era)

class Indexable elem container where

Methods

indexOf ∷ AsItem Word32 elem → container → StrictMaybe (AsIx Word32 elem)

fromIndex ∷ AsIx Word32 elem → container → StrictMaybe (AsIxItem Word32 elem)

inputs' ∷ TxBody AlonzoEra → Set TxIn

Deprecated: In favor of inputsTxBodyL

collateral' ∷ TxBody AlonzoEra → Set TxIn

Deprecated: In favor of collateralInputsTxBodyL

outputs' ∷ TxBody AlonzoEra → StrictSeq (TxOut AlonzoEra)

Deprecated: In favor of outputsTxBodyL

certs' ∷ TxBody AlonzoEra → StrictSeq (TxCert AlonzoEra)

Deprecated: In favor of certsTxBodyL

withdrawals' ∷ TxBody AlonzoEra → Withdrawals

Deprecated: In favor of withdrawalsTxBodyL

txfee' ∷ TxBody AlonzoEra → Coin

Deprecated: In favor of feeTxBodyL

vldt' ∷ TxBody AlonzoEra → ValidityInterval

Deprecated: In favor of vldtTxBodyL

update' ∷ TxBody AlonzoEra → StrictMaybe (Update AlonzoEra)

Deprecated: In favor of updateTxBodyL

reqSignerHashes' ∷ TxBody AlonzoEra → Set (KeyHash 'Witness)

Deprecated: In favor of reqSignerHashesTxBodyL

mint' ∷ TxBody AlonzoEra → MultiAsset

Deprecated: In favor of mintTxBodyL

scriptIntegrityHash' ∷ TxBody AlonzoEra → StrictMaybe ScriptIntegrityHash

Deprecated: In favor of scriptIntegrityHashTxBodyL

adHash' ∷ TxBody AlonzoEra → StrictMaybe TxAuxDataHash

Deprecated: In favor of auxDataHashTxBodyL

txnetworkid' ∷ TxBody AlonzoEra → StrictMaybe Network

Deprecated: In favor of networkIdTxBodyL

getAdaOnly ∷ ∀ era. Val (Value era) ⇒ Proxy era → Value era → Maybe (CompactForm Coin)

decodeDataHash32 ∷ DataHash32 → DataHash

encodeDataHash32 ∷ DataHash → DataHash32

encodeAddress28 ∷ Network → PaymentCredential → Addr28Extra

decodeAddress28 ∷ Credential 'Staking → Addr28Extra → Addr

viewCompactTxOut ∷ Val (Value era) ⇒ AlonzoTxOut era → (CompactAddr, CompactForm (Value era), StrictMaybe DataHash)

viewTxOut ∷ Val (Value era) ⇒ AlonzoTxOut era → (Addr, Value era, StrictMaybe DataHash)

data EraIndependentScriptIntegrity

type ScriptIntegrityHash = SafeHash EraIndependentScriptIntegrity

getAlonzoTxOutEitherAddr ∷ AlonzoTxOut era → Either Addr CompactAddr

utxoEntrySize ∷ AlonzoEraTxOut era ⇒ TxOut era → Integer

Compute an estimate of the size of storing one UTxO entry. This function implements the UTxO entry size estimate done by scaledMinDeposit in the ShelleyMA era

alonzoRedeemerPointer ∷ ∀ era. MaryEraTxBody era ⇒ TxBody era → AlonzoPlutusPurpose AsItem era → StrictMaybe (AlonzoPlutusPurpose AsIx era)

alonzoRedeemerPointerInverse ∷ MaryEraTxBody era ⇒ TxBody era → AlonzoPlutusPurpose AsIx era → StrictMaybe (AlonzoPlutusPurpose AsIxItem era)

Orphan instances

AllegraEraTxBody AlonzoEra

Instance details

Methods

vldtTxBodyL ∷ Lens' (TxBody AlonzoEra) ValidityInterval

EraTxBody AlonzoEra

Instance details

Associated Types

data TxBody AlonzoEra

Methods

mkBasicTxBody ∷ TxBody AlonzoEra

inputsTxBodyL ∷ Lens' (TxBody AlonzoEra) (Set TxIn)

outputsTxBodyL ∷ Lens' (TxBody AlonzoEra) (StrictSeq (TxOut AlonzoEra))

feeTxBodyL ∷ Lens' (TxBody AlonzoEra) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody AlonzoEra) Withdrawals

auxDataHashTxBodyL ∷ Lens' (TxBody AlonzoEra) (StrictMaybe TxAuxDataHash)

spendableInputsTxBodyF ∷ SimpleGetter (TxBody AlonzoEra) (Set TxIn)

allInputsTxBodyF ∷ SimpleGetter (TxBody AlonzoEra) (Set TxIn)

certsTxBodyL ∷ Lens' (TxBody AlonzoEra) (StrictSeq (TxCert AlonzoEra))

getTotalDepositsTxBody ∷ PParams AlonzoEra → (KeyHash 'StakePool → Bool) → TxBody AlonzoEra → Coin

getTotalRefundsTxBody ∷ PParams AlonzoEra → (Credential 'Staking → Maybe Coin) → (Credential 'DRepRole → Maybe Coin) → TxBody AlonzoEra → Coin

getGenesisKeyHashCountTxBody ∷ TxBody AlonzoEra → Int

MaryEraTxBody AlonzoEra

Instance details

Methods

mintTxBodyL ∷ Lens' (TxBody AlonzoEra) MultiAsset

mintedTxBodyF ∷ SimpleGetter (TxBody AlonzoEra) (Set PolicyID)

mintValueTxBodyF ∷ SimpleGetter (TxBody AlonzoEra) (Value AlonzoEra)

ShelleyEraTxBody AlonzoEra

Instance details

Methods

ttlTxBodyL ∷ Lens' (TxBody AlonzoEra) SlotNo

updateTxBodyL ∷ Lens' (TxBody AlonzoEra) (StrictMaybe (Update AlonzoEra))

Generic (TxBody AlonzoEra)

Instance details

Associated Types

type Rep (TxBody AlonzoEra) ∷ Type → Type

Methods

from ∷ TxBody AlonzoEra → Rep (TxBody AlonzoEra) x

to ∷ Rep (TxBody AlonzoEra) x → TxBody AlonzoEra

Show (TxBody AlonzoEra)

Instance details

Methods

showsPrec ∷ Int → TxBody AlonzoEra → ShowS

show ∷ TxBody AlonzoEra → String

showList ∷ [TxBody AlonzoEra] → ShowS

ToCBOR (TxBody AlonzoEra)

Instance details

Methods

toCBOR ∷ TxBody AlonzoEra → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy (TxBody AlonzoEra) → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [TxBody AlonzoEra] → Size

DecCBOR (Annotator (TxBody AlonzoEra))

Instance details

Methods

decCBOR ∷ Decoder s (Annotator (TxBody AlonzoEra))

dropCBOR ∷ Proxy (Annotator (TxBody AlonzoEra)) → Decoder s ()

label ∷ Proxy (Annotator (TxBody AlonzoEra)) → Text

EncCBOR (TxBody AlonzoEra)
Encodes memoized bytes created upon construction.

Instance details

Methods

encCBOR ∷ TxBody AlonzoEra → Encoding

encodedSizeExpr ∷ (∀ t. EncCBOR t ⇒ Proxy t → Size) → Proxy (TxBody AlonzoEra) → Size

encodedListSizeExpr ∷ (∀ t. EncCBOR t ⇒ Proxy t → Size) → Proxy [TxBody AlonzoEra] → Size

SafeToHash (TxBody AlonzoEra)

Instance details

Methods

originalBytes ∷ TxBody AlonzoEra → ByteString

originalBytesSize ∷ TxBody AlonzoEra → Int

makeHashWithExplicitProxys ∷ Proxy i → TxBody AlonzoEra → SafeHash i

EqRaw (TxBody AlonzoEra)

Instance details

Methods

eqRaw ∷ TxBody AlonzoEra → TxBody AlonzoEra → Bool

Memoized (TxBody AlonzoEra)

Instance details

Associated Types

type RawType (TxBody AlonzoEra) = (r ∷ Type)

Methods

getMemoBytes ∷ TxBody AlonzoEra → MemoBytes (RawType (TxBody AlonzoEra))

wrapMemoBytes ∷ MemoBytes (RawType (TxBody AlonzoEra)) → TxBody AlonzoEra

NFData (TxBody AlonzoEra)

Instance details

Methods

rnf ∷ TxBody AlonzoEra → ()

Eq (TxBody AlonzoEra)

Instance details

Methods

(==) ∷ TxBody AlonzoEra → TxBody AlonzoEra → Bool

(/=) ∷ TxBody AlonzoEra → TxBody AlonzoEra → Bool

NoThunks (TxBody AlonzoEra)

Instance details

Methods

noThunks ∷ Context → TxBody AlonzoEra → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → TxBody AlonzoEra → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (TxBody AlonzoEra) → String

HashAnnotated (TxBody AlonzoEra) EraIndependentTxBody

Instance details

Methods

hashAnnotated ∷ TxBody AlonzoEra → SafeHash EraIndependentTxBody