Cardano.Ledger.Alonzo.TxInfo

Documentation

data TxOutSource c

A transaction output can be translated because it is a newly created output, or because it is the output which is connected to a transaction input being spent.

Constructors

TxOutFromInput !(TxIn c)

TxOutFromOutput !TxIx

data TranslationError c

NOTE: class TranslateEra defines an associated type with the same name. Not to be confused.

Constructors

ByronTxOutInContext !(TxOutSource c)

TranslationLogicMissingInput !(TxIn c)

RdmrPtrPointsToNothing !RdmrPtr

LanguageNotSupported !Language

InlineDatumsNotSupported !(TxOutSource c)

ReferenceScriptsNotSupported !(TxOutSource c)

ReferenceInputsNotSupported !(Set (TxIn c))

TimeTranslationPastHorizon !Text

transProtocolVersion ∷ ProtVer → ProtocolVersion

validScript ∷ ProtVer → AlonzoScript era → Bool

Test that every Alonzo script represents a real Script. Run deepseq to see that there are no infinite computations and that every Plutus Script unflattens into a real PV1.Script

transDataHash ∷ StrictMaybe (DataHash c) → Maybe DatumHash

transDataHash' ∷ DataHash c → DatumHash

transKeyHash ∷ KeyHash d c → PubKeyHash

transSafeHash ∷ SafeHash c i → BuiltinByteString

transHash ∷ Hash h a → ByteString

txInfoId ∷ TxId c → TxId

transStakeCred ∷ Credential kr c → Credential

transStakeReference ∷ StakeReference c → Maybe StakingCredential

transCred ∷ Credential kr c → Credential

transAddr ∷ Addr c → Maybe Address

transTxOutAddr ∷ EraTxOut era ⇒ TxOut era → Maybe Address

slotToPOSIXTime ∷ EpochInfo (Either Text) → SystemStart → SlotNo → Either Text POSIXTime

transVITime ∷ EraPParams era ⇒ PParams era → EpochInfo (Either Text) → SystemStart → ValidityInterval → Either Text POSIXTimeRange

translate a validity interval to POSIX time

txInfoIn' ∷ TxIn c → TxOutRef

txInfoIn ∷ (AlonzoEraTxOut era, Value era ~ MaryValue (EraCrypto era)) ⇒ TxIn (EraCrypto era) → TxOut era → Maybe TxInInfo

Given a TxIn, look it up in the UTxO. If it exists, translate it and return (Just translation). If does not exist in the UTxO, return Nothing.

txInfoOut ∷ (AlonzoEraTxOut era, Value era ~ MaryValue (EraCrypto era)) ⇒ TxOut era → Maybe TxOut

Given a TxOut, translate it and return (Just transalation). It is possible the address part is a Bootstrap Address, in that case return Nothing I.e. don't include Bootstrap Addresses in the answer.

transPolicyID ∷ PolicyID c → CurrencySymbol

transAssetName ∷ AssetName → TokenName

transMultiAsset ∷ MultiAsset c → Value

transMintValue ∷ MultiAsset c → Value

Hysterical raisins:

Previously transaction body contained a mint field with MaryValue instead of a MultiAsset, which has changed since then to just MultiAsset (because minting ADA makes no sense). However, if we don't preserve previous translation, scripts that previously succeeded will fail.

transValue ∷ MaryValue c → Value

transDCert ∷ DCert c → DCert

transWithdrawals ∷ Withdrawals c → Map StakingCredential Integer

getWitVKeyHash ∷ (Crypto c, Typeable kr) ⇒ WitVKey kr c → PubKeyHash

transDataPair ∷ (DataHash c, Data era) → (DatumHash, Datum)

transExUnits ∷ ExUnits → ExBudget

exBudgetToExUnits ∷ ExBudget → Maybe ExUnits

transScriptPurpose ∷ ScriptPurpose c → ScriptPurpose

data VersionedTxInfo

Constructors

TxInfoPV1 TxInfo

TxInfoPV2 TxInfo

class ExtendedUTxO era where

Where we keep functions that differ from Era to Era but which deal with the extra things in the TxOut (Scripts, DataHash, Datum, etc)

Methods

txInfo ∷ PParams era → Language → EpochInfo (Either Text) → SystemStart → UTxO era → Tx era → Either (TranslationError (EraCrypto era)) VersionedTxInfo

txscripts ∷ UTxO era → Tx era → Map (ScriptHash (EraCrypto era)) (Script era)

getAllowedSupplimentalDataHashes ∷ TxBody era → UTxO era → Set (DataHash (EraCrypto era))

getDatum ∷ Tx era → UTxO era → ScriptPurpose (EraCrypto era) → Maybe (Data era)

alonzoTxInfo ∷ ∀ era. (EraTx era, AlonzoEraTxBody era, Value era ~ MaryValue (EraCrypto era), TxWits era ~ AlonzoTxWits era) ⇒ PParams era → Language → EpochInfo (Either Text) → SystemStart → UTxO era → Tx era → Either (TranslationError (EraCrypto era)) VersionedTxInfo

valContext ∷ Era era ⇒ VersionedTxInfo → ScriptPurpose (EraCrypto era) → Data era

valContext pairs transaction data with a script purpose. See figure 22 of the Alonzo specification.

data ScriptFailure

Constructors

PlutusSF Text PlutusDebug

data ScriptResult

Constructors

Passes [PlutusDebug]

Fails [PlutusDebug] (NonEmpty ScriptFailure)

scriptPass ∷ PlutusDebug → ScriptResult

scriptFail ∷ ScriptFailure → ScriptResult

data PlutusDebugLang (l ∷ Language) where

Constructors

PlutusDebugLang

Fields

∷ { pdSLanguage ∷ SLanguage l

, pdCostModel ∷ CostModel

, pdExUnits ∷ ExUnits

, pdSBS ∷ ShortByteString

, pdPlutusData ∷ PlutusData

, pdProtVer ∷ ProtVer

} → PlutusDebugLang l

data PlutusDebug where

Constructors

PlutusDebug ∷ IsLanguage l ⇒ PlutusDebugLang l → PlutusDebug

newtype PlutusData

Constructors

PlutusData

Fields

unPlutusData ∷ [Data]

data PlutusError

Constructors

PlutusErrorV1 EvaluationError

PlutusErrorV2 EvaluationError

data PlutusDebugInfo

Constructors

DebugSuccess ExBudget

DebugCannotDecode String

DebugInfo [Text] PlutusError PlutusDebug

DebugBadHex String

debugPlutus ∷ Version → String → PlutusDebugInfo

runPLCScript ∷ ∀ era. Show (AlonzoScript era) ⇒ Proxy era → ProtVer → Language → CostModel → ShortByteString → ExUnits → [Data] → ScriptResult

Run a Plutus Script, given the script and the bounds on resources it is allocated.

explainPlutusFailure ∷ ∀ era. Show (AlonzoScript era) ⇒ Proxy era → ProtVer → Language → ShortByteString → EvaluationError → [Data] → CostModel → ExUnits → ScriptResult

Explain why a script might fail. Scripts come in two flavors:

with 3 data arguments [data,redeemer,context]

with 2 data arguments [redeemer,context].

It pays to decode the context data into a real context because that provides way more information. But there is no guarantee the context data really can be decoded.

languages ∷ ∀ era. (ExtendedUTxO era, Script era ~ AlonzoScript era) ⇒ Tx era → UTxO era → Set (ScriptHash (EraCrypto era)) → Set Language

Compute the Set of Languages in an era, where AlonzoScripts are used

validPlutusdata ∷ Data → Bool

Deprecated: Plutus data bytestrings are not restricted to sixty-four bytes.

getTxOutDatum ∷ AlonzoEraTxOut era ⇒ TxOut era → Datum era

Deprecated: In favor of datumTxOutF