Cardano.Ledger.Alonzo.TxSeq

Description

TxSeq. This is effectively the block body, which consists of a sequence of transactions with segregated witness and metadata information.

Documentation

data AlonzoTxSeq era where

Bundled Patterns

pattern AlonzoTxSeq ∷ ∀ era. (AlonzoEraTx era, SafeToHash (TxWits era)) ⇒ StrictSeq (Tx era) → AlonzoTxSeq era

type TxSeq era = AlonzoTxSeq era

Deprecated: Use AlonzoTxSeq instead

hashTxSeq ∷ ∀ era. Era era ⇒ AlonzoTxSeq era → Hash (EraCrypto era) EraIndependentBlockBody

Deprecated: Use hashAlonzoTxSeq instead

hashAlonzoTxSeq ∷ ∀ era. Era era ⇒ AlonzoTxSeq era → Hash (EraCrypto era) EraIndependentBlockBody

Hash a given block body