Cardano.Ledger.Alonzo.TxSeq

Description

TxSeq. This is effectively the block body, which consists of a sequence of transactions with segregated witness and metadata information.

Documentation

data TxSeq era where

Bundled Patterns

pattern TxSeq :: forall era. (Era era, SafeToHash (TxWitness era)) => StrictSeq (ValidatedTx era) -> TxSeq era

hashTxSeq :: forall era. Era era => TxSeq era -> Hash (Crypto era) EraIndependentBlockBody

Hash a given block body