Cardano.Ledger.Alonzo.TxWits

Documentation

newtype Redeemers era

Note that Redeemers are based on MemoBytes since we must preserve the original bytes for the ScriptIntegrity. Since the Redeemers exist outside of the transaction body, this is how we ensure that they are not manipulated.

Constructors

MkRedeemers (MemoBytes (RedeemersRaw era))

Bundled Patterns

pattern Redeemers ∷ ∀ era. AlonzoEraScript era ⇒ Map (PlutusPurpose AsIx era) (Data era, ExUnits) → Redeemers era

newtype RedeemersRaw era

Constructors

RedeemersRaw

Fields

unRedeemersRaw ∷ Map (PlutusPurpose AsIx era) (Data era, ExUnits)

unRedeemersL ∷ ∀ era. AlonzoEraScript era ⇒ Lens' (Redeemers era) (Map (PlutusPurpose AsIx era) (Data era, ExUnits))

unRedeemers ∷ Redeemers era → Map (PlutusPurpose AsIx era) (Data era, ExUnits)

nullRedeemers ∷ Redeemers era → Bool

Deprecated: In favor of unRedeemersL

lookupRedeemer ∷ Ord (PlutusPurpose AsIx era) ⇒ PlutusPurpose AsIx era → Redeemers era → Maybe (Data era, ExUnits)

Deprecated: In favor of unRedeemersL

newtype TxDats era

Note that TxDats are based on MemoBytes since we must preserve the original bytes for the ScriptIntegrity. Since the TxDats exist outside of the transaction body, this is how we ensure that they are not manipulated.

Constructors

MkTxDats (MemoBytes (TxDatsRaw era))

Bundled Patterns

pattern TxDats ∷ ∀ era. Era era ⇒ Map DataHash (Data era) → TxDats era

pattern TxDats' ∷ Map DataHash (Data era) → TxDats era

Deprecated: In favor of TxDats

newtype TxDatsRaw era

Constructors

TxDatsRaw

Fields

unTxDatsRaw ∷ Map DataHash (Data era)

newtype AlonzoTxWits era

Constructors

MkAlonzoTxWits (MemoBytes (AlonzoTxWitsRaw era))

Bundled Patterns

pattern AlonzoTxWits ∷ ∀ era. AlonzoEraScript era ⇒ Set (WitVKey 'Witness) → Set BootstrapWitness → Map ScriptHash (Script era) → TxDats era → Redeemers era → AlonzoTxWits era

pattern AlonzoTxWits' ∷ Era era ⇒ Set (WitVKey 'Witness) → Set BootstrapWitness → Map ScriptHash (Script era) → TxDats era → Redeemers era → AlonzoTxWits era

Deprecated: In favor of AlonzoTxWits

data AlonzoTxWitsRaw era

Internal AlonzoTxWits type, lacking serialised bytes.

Constructors

AlonzoTxWitsRaw

Fields

atwrAddrTxWits ∷ !(Set (WitVKey 'Witness))

atwrBootAddrTxWits ∷ !(Set BootstrapWitness)

atwrScriptTxWits ∷ !(Map ScriptHash (Script era))

atwrDatsTxWits ∷ !(TxDats era)

atwrRdmrsTxWits ∷ !(Redeemers era)

addrAlonzoTxWitsL ∷ ∀ era. AlonzoEraScript era ⇒ Lens' (AlonzoTxWits era) (Set (WitVKey 'Witness))

bootAddrAlonzoTxWitsL ∷ ∀ era. AlonzoEraScript era ⇒ Lens' (AlonzoTxWits era) (Set BootstrapWitness)

scriptAlonzoTxWitsL ∷ ∀ era. AlonzoEraScript era ⇒ Lens' (AlonzoTxWits era) (Map ScriptHash (Script era))

datsAlonzoTxWitsL ∷ ∀ era. AlonzoEraScript era ⇒ Lens' (AlonzoTxWits era) (TxDats era)

rdmrsAlonzoTxWitsL ∷ ∀ era. AlonzoEraScript era ⇒ Lens' (AlonzoTxWits era) (Redeemers era)

class (EraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTxWits era where

Methods

datsTxWitsL ∷ Lens' (TxWits era) (TxDats era)

rdmrsTxWitsL ∷ Lens' (TxWits era) (Redeemers era)

hashDataTxWitsL ∷ AlonzoEraTxWits era ⇒ Lens (TxWits era) (TxWits era) (TxDats era) [Data era]

This is a convenience Lens that will hash the Data when it is being added to the TxWits. See datsTxWitsL for a version that aloows setting TxDats instead.

unTxDats ∷ TxDats era → Map DataHash (Data era)

unTxDatsL ∷ ∀ era. Era era ⇒ Lens' (TxDats era) (Map DataHash (Data era))

nullDats ∷ TxDats era → Bool

Deprecated: In favor of unTxDatsL

alonzoEqTxWitsRaw ∷ AlonzoEraTxWits era ⇒ TxWits era → TxWits era → Bool

emptyTxWitsRaw ∷ AlonzoEraScript era ⇒ AlonzoTxWitsRaw era

addScriptsTxWitsRaw ∷ Map ScriptHash (Script era) → AlonzoTxWitsRaw era → AlonzoTxWitsRaw era

decodeAlonzoPlutusScript ∷ (AlonzoEraScript era, PlutusLanguage l) ⇒ SLanguage l → Decode ('Closed 'Dense) (Map ScriptHash (Script era))

asHashedScriptPair ∷ ∀ era. EraScript era ⇒ Script era → (ScriptHash, Script era)

Orphan instances

EraScript AlonzoEra ⇒ EraTxWits AlonzoEra

Instance details

Associated Types

type TxWits AlonzoEra = (r ∷ Type)

Methods

mkBasicTxWits ∷ TxWits AlonzoEra

addrTxWitsL ∷ Lens' (TxWits AlonzoEra) (Set (WitVKey 'Witness))

bootAddrTxWitsL ∷ Lens' (TxWits AlonzoEra) (Set BootstrapWitness)

scriptTxWitsL ∷ Lens' (TxWits AlonzoEra) (Map ScriptHash (Script AlonzoEra))