Cardano.Ledger.Alonzo.TxWits

Documentation

data Redeemers era where

Note that Redeemers are based on MemoBytes since we must preserve the original bytes for the ScriptIntegrity. Since the Redeemers exist outside of the transaction body, this is how we ensure that they are not manipulated.

Bundled Patterns

pattern Redeemers ∷ ∀ era. AlonzoEraScript era ⇒ Map (PlutusPurpose AsIx era) (Data era, ExUnits) → Redeemers era

data RedeemersRaw era

unRedeemers ∷ Redeemers era → Map (PlutusPurpose AsIx era) (Data era, ExUnits)

nullRedeemers ∷ Redeemers era → Bool

lookupRedeemer ∷ Ord (PlutusPurpose AsIx era) ⇒ PlutusPurpose AsIx era → Redeemers era → Maybe (Data era, ExUnits)

upgradeRedeemers ∷ ∀ era. (AlonzoEraScript (PreviousEra era), AlonzoEraScript era) ⇒ Redeemers (PreviousEra era) → Redeemers era

Upgrade redeemers from one era to another. The underlying data structure will remain identical, but the memoised serialisation may change to reflect the versioned serialisation of the new era.

data TxDats era where

Note that TxDats are based on MemoBytes since we must preserve the original bytes for the ScriptIntegrity. Since the TxDats exist outside of the transaction body, this is how we ensure that they are not manipulated.

Bundled Patterns

pattern TxDats ∷ Era era ⇒ Map (DataHash (EraCrypto era)) (Data era) → TxDats era

pattern TxDats' ∷ Map (DataHash (EraCrypto era)) (Data era) → TxDats era

data TxDatsRaw era

upgradeTxDats ∷ (Era era1, Era era2, EraCrypto era1 ~ EraCrypto era2) ⇒ TxDats era1 → TxDats era2

Upgrade TxDats from one era to another. The underlying data structure will remain identical, but the memoised serialisation may change to reflect the versioned serialisation of the new era.

data AlonzoTxWits era where

Bundled Patterns

pattern AlonzoTxWits ∷ AlonzoEraScript era ⇒ Set (WitVKey 'Witness (EraCrypto era)) → Set (BootstrapWitness (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (Script era) → TxDats era → Redeemers era → AlonzoTxWits era

pattern AlonzoTxWits' ∷ Era era ⇒ Set (WitVKey 'Witness (EraCrypto era)) → Set (BootstrapWitness (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (Script era) → TxDats era → Redeemers era → AlonzoTxWits era

data AlonzoTxWitsRaw era

Internal AlonzoTxWits type, lacking serialised bytes.

addrAlonzoTxWitsL ∷ AlonzoEraScript era ⇒ Lens' (AlonzoTxWits era) (Set (WitVKey 'Witness (EraCrypto era)))

bootAddrAlonzoTxWitsL ∷ AlonzoEraScript era ⇒ Lens' (AlonzoTxWits era) (Set (BootstrapWitness (EraCrypto era)))

scriptAlonzoTxWitsL ∷ AlonzoEraScript era ⇒ Lens' (AlonzoTxWits era) (Map (ScriptHash (EraCrypto era)) (Script era))

datsAlonzoTxWitsL ∷ AlonzoEraScript era ⇒ Lens' (AlonzoTxWits era) (TxDats era)

rdmrsAlonzoTxWitsL ∷ AlonzoEraScript era ⇒ Lens' (AlonzoTxWits era) (Redeemers era)

class (EraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTxWits era where

Methods

datsTxWitsL ∷ Lens' (TxWits era) (TxDats era)

rdmrsTxWitsL ∷ Lens' (TxWits era) (Redeemers era)

hashDataTxWitsL ∷ AlonzoEraTxWits era ⇒ Lens (TxWits era) (TxWits era) (TxDats era) [Data era]

This is a convenience Lens that will hash the Data when it is being added to the TxWits. See datsTxWitsL for a version that aloows setting TxDats instead.

unTxDats ∷ TxDats era → Map (DataHash (EraCrypto era)) (Data era)

nullDats ∷ TxDats era → Bool

alonzoEqTxWitsRaw ∷ AlonzoEraTxWits era ⇒ TxWits era → TxWits era → Bool

Orphan instances

(EraScript (AlonzoEra c), Crypto c) ⇒ EraTxWits (AlonzoEra c)

Instance details

Associated Types

type TxWits (AlonzoEra c) = (r ∷ Type)

Methods

mkBasicTxWits ∷ TxWits (AlonzoEra c)

addrTxWitsL ∷ Lens' (TxWits (AlonzoEra c)) (Set (WitVKey 'Witness (EraCrypto (AlonzoEra c))))

bootAddrTxWitsL ∷ Lens' (TxWits (AlonzoEra c)) (Set (BootstrapWitness (EraCrypto (AlonzoEra c))))

scriptTxWitsL ∷ Lens' (TxWits (AlonzoEra c)) (Map (ScriptHash (EraCrypto (AlonzoEra c))) (Script (AlonzoEra c)))

upgradeTxWits ∷ TxWits (PreviousEra (AlonzoEra c)) → TxWits (AlonzoEra c)