Cardano.Ledger.Alonzo.TxWits

Documentation

data RdmrPtr

Constructors

RdmrPtr !Tag !Word64

data Redeemers era where

Note that Redeemers are based on MemoBytes since we must preserve the original bytes for the ScriptIntegrity. Since the Redeemers exist outside of the transaction body, this is how we ensure that they are not manipulated.

Bundled Patterns

pattern Redeemers ∷ ∀ era. Era era ⇒ Map RdmrPtr (Data era, ExUnits) → Redeemers era

unRedeemers ∷ Era era ⇒ Redeemers era → Map RdmrPtr (Data era, ExUnits)

nullRedeemers ∷ Era era ⇒ Redeemers era → Bool

data TxDats era where

Note that TxDats are based on MemoBytes since we must preserve the original bytes for the ScriptIntegrity. Since the TxDats exist outside of the transaction body, this is how we ensure that they are not manipulated.

Bundled Patterns

pattern TxDats ∷ Era era ⇒ Map (DataHash (EraCrypto era)) (Data era) → TxDats era

pattern TxDats' ∷ Map (DataHash (EraCrypto era)) (Data era) → TxDats era

data AlonzoTxWits era where

Bundled Patterns

pattern AlonzoTxWits ∷ (Era era, Script era ~ AlonzoScript era) ⇒ Set (WitVKey 'Witness (EraCrypto era)) → Set (BootstrapWitness (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (Script era) → TxDats era → Redeemers era → AlonzoTxWits era

pattern AlonzoTxWits' ∷ Era era ⇒ Set (WitVKey 'Witness (EraCrypto era)) → Set (BootstrapWitness (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (Script era) → TxDats era → Redeemers era → AlonzoTxWits era

addrAlonzoTxWitsL ∷ (Era era, Script era ~ AlonzoScript era) ⇒ Lens' (AlonzoTxWits era) (Set (WitVKey 'Witness (EraCrypto era)))

bootAddrAlonzoTxWitsL ∷ (Era era, Script era ~ AlonzoScript era) ⇒ Lens' (AlonzoTxWits era) (Set (BootstrapWitness (EraCrypto era)))

scriptAlonzoTxWitsL ∷ (Era era, Script era ~ AlonzoScript era) ⇒ Lens' (AlonzoTxWits era) (Map (ScriptHash (EraCrypto era)) (Script era))

datsAlonzoTxWitsL ∷ (Era era, Script era ~ AlonzoScript era) ⇒ Lens' (AlonzoTxWits era) (TxDats era)

rdmrsAlonzoTxWitsL ∷ (Era era, Script era ~ AlonzoScript era) ⇒ Lens' (AlonzoTxWits era) (Redeemers era)

class EraTxWits era ⇒ AlonzoEraTxWits era where

Methods

datsTxWitsL ∷ Lens' (TxWits era) (TxDats era)

rdmrsTxWitsL ∷ Lens' (TxWits era) (Redeemers era)

hashDataTxWitsL ∷ AlonzoEraTxWits era ⇒ Lens (TxWits era) (TxWits era) (TxDats era) [Data era]

This is a convenience Lens that will hash the Data when it is being added to the TxWits. See datsTxWitsL for a version that aloows setting TxDats instead.

unTxDats ∷ TxDats era → Map (DataHash (EraCrypto era)) (Data era)

nullDats ∷ TxDats era → Bool

Orphan instances

(EraScript (AlonzoEra c), Crypto c) ⇒ EraTxWits (AlonzoEra c)

Instance details

Associated Types

type TxWits (AlonzoEra c) = (r ∷ Type)

Methods

mkBasicTxWits ∷ TxWits (AlonzoEra c)

addrTxWitsL ∷ Lens' (TxWits (AlonzoEra c)) (Set (WitVKey 'Witness (EraCrypto (AlonzoEra c))))

bootAddrTxWitsL ∷ Lens' (TxWits (AlonzoEra c)) (Set (BootstrapWitness (EraCrypto (AlonzoEra c))))

scriptTxWitsL ∷ Lens' (TxWits (AlonzoEra c)) (Map (ScriptHash (EraCrypto (AlonzoEra c))) (Script (AlonzoEra c)))